\name{as.data.frame.testhat_results}
\alias{as.data.frame.testhat_results}
\title{convert testthat_results to data frame}
\usage{
  as.data.frame.testhat_results(x, ...)
}
\arguments{
  \item{x}{the testthat results}

  \item{...}{ignored}
}
\value{
  a data frame
}
\description{
  convert testthat_results to data frame
}

