\name{auto_test_package}
\alias{auto_test_package}
\title{Watches a package for changes, rerunning tests as appropriate.}
\usage{auto_test_package(path, suite=SummarySuite)}

\description{
  Watches a package for changes, rerunning tests as appropriate.
}
\keyword{debugging}
\seealso{\code{\link{auto_test}} for details on how method works}
\arguments{
  \item{path}{path to package}
  \item{suite}{test reporter to use}
}
