% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernie.R
\docType{data}
\name{bernie}
\alias{bernie}
\title{Bernie}
\format{
A list with two elements
\describe{
\item{point}{Bernie is pointing}
\item{chair}{Bernie is in a chair}
}
}
\source{
Pointing bernie from https://www.pngfind.com/mpng/TRixTi_create-your-pic-bernie-sanders-head-png-transparent/, Bernie in chair from https://twitter.com/taber/status/1351938194024124419 and both turned to ASCII in https://manytools.org/hacker-tools/convert-images-to-ascii-art
}
\usage{
bernie
}
\description{
ASCII art of Bernie Sanders
}
\keyword{datasets}
