% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{browser2}
\alias{browser2}
\title{Call \code{browser()} with sinks temporarily diverted to \code{stdout}}
\usage{
browser2(..., skipCalls = 0, frame = parent.frame())
}
\arguments{
\item{..., skipCalls}{Arguments passed to \code{browser()}.}

\item{frame}{The execution environment in which to register an
\code{on.exit} expression to restore any existing sinks.}
}
\description{
\code{browser2()} is useful to browse through code when an output sink
like \code{testthat::expect_snapshot()} is on the stack. The sinks are
restored to avoid disturbing the snapshot tests. Of course you will
still get failures if you're browsing through any \code{cat()} code (or
equivalent) that should be captured by a snapshot.
}
