% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock-object.R
\name{call-expectations}
\alias{call-expectations}
\alias{expect_args}
\alias{expect_call}
\title{Expectation: does the given call match the expected?}
\usage{
expect_call(mock_object, n, expected_call)

expect_args(mock_object, n, ...)
}
\arguments{
\item{mock_object}{A \code{\link{mock}} object.}

\item{n}{Call number.}

\item{expected_call}{Expected call expression; will be compared unevaluated.}

\item{...}{Arguments as passed in a call.}
}
\description{
Together with \code{\link{mock}} can be used to verify whether the
call expression (\code{\link{expect_call}}) and/or argument values
(\code{\link{expect_args}}) match the expected.
}
\examples{
# expect call expression (signature)
m <- mock()
with_mock(summary = m, summary(iris))
expect_call(m, 1, summary(iris))

# expect argument value
m <- mock()
a <- iris
with_mock(summary = m, summary(object = a))
expect_args(m, 1, object = a)
# is an equivalent to ...
expect_equal(mock_args(m)[[1]], list(object = a))

}
\seealso{
Other expectations: \code{\link{comparison-expectations}},
  \code{\link{equality-expectations}},
  \code{\link{expect_equal_to_reference}},
  \code{\link{expect_length}}, \code{\link{expect_match}},
  \code{\link{expect_named}},
  \code{\link{inheritance-expectations}},
  \code{\link{logical-expectations}},
  \code{\link{output-expectations}}
}

