% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-check.R
\name{CheckReporter}
\alias{CheckReporter}
\title{Report results for \verb{R CMD check}}
\description{
\verb{R CMD check} displays only the last 13 lines of the result, so this
report is designed to ensure that you see something useful there.
}
\seealso{
Other reporters: 
\code{\link{DebugReporter}},
\code{\link{FailReporter}},
\code{\link{JunitReporter}},
\code{\link{ListReporter}},
\code{\link{LlmReporter}},
\code{\link{LocationReporter}},
\code{\link{MinimalReporter}},
\code{\link{MultiReporter}},
\code{\link{ProgressReporter}},
\code{\link{RStudioReporter}},
\code{\link{Reporter}},
\code{\link{SilentReporter}},
\code{\link{SlowReporter}},
\code{\link{SonarqubeReporter}},
\code{\link{StopReporter}},
\code{\link{SummaryReporter}},
\code{\link{TapReporter}},
\code{\link{TeamcityReporter}}
}
\concept{reporters}
