% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect-is.R
\name{class-expectations}
\alias{class-expectations}
\alias{expect_is}
\alias{expect_null}
\alias{expect_type}
\title{Expectation: does the object inherit from a S3 or S4 class, or a base type?}
\usage{
expect_is(object, class, info = NULL, label = NULL)

expect_type(object, type)

expect_null(object, info = NULL, label = NULL)
}
\arguments{
\item{object}{object to test}

\item{class}{character vector of class names}

\item{info}{extra information to be included in the message (useful when
writing tests in loops).}

\item{label}{object label. When \code{NULL}, computed from deparsed object.}

\item{type}{String giving base type (as returned by \code{\link{typeof}}).}
}
\description{
Tests whether or not an object inherits from any of a list of classes, or
is an instance of a base type. \code{expect_null} is a special case designed
for detecting \code{NULL}.
}
\examples{
expect_is(1, "numeric")
a <- matrix(1:10, nrow = 5)
expect_is(a, "matrix")

expect_is(mtcars, "data.frame")
# alternatively for classes that have an is method
expect_true(is.data.frame(mtcars))

f <- factor("a")
expect_is(f, "factor")
expect_type(f, "integer")

expect_null(NULL)

}
\seealso{
\code{\link{inherits}}

Other expectations: \code{\link{comparison-expectations}},
  \code{\link{equality-expectations}},
  \code{\link{expect_equal_to_reference}},
  \code{\link{expect_match}}, \code{\link{expect_named}},
  \code{\link{logical-expectations}},
  \code{\link{output-expectations}},
  \code{\link{takes_less_than}}
}

