% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compare.r
\name{compare}
\alias{compare}
\alias{compare.character}
\alias{compare.default}
\title{Provide human-readable comparison of two objects}
\usage{
compare(x, y, ...)

\method{compare}{default}(x, y, ...)

\method{compare}{character}(x, y, ..., max_strings = 5, max_lines = 5,
  width = getOption("width"))
}
\arguments{
\item{x,y}{Objects to compare}

\item{...}{Additional arguments used to control specifics of comparison}

\item{max_strings}{Maximum number of differences to show}

\item{max_lines}{Maximum number of lines to show from each difference}

\item{width}{Width of output device}
}
\description{
\code{compare} is similar to \code{\link[base]{all.equal}()}, but is
tailored for unit test output. In fact, the default method is to just
use \code{all.equal}.
}
\examples{
x <- c("abc", "def", "jih")
compare(x, x)

y <- paste0(x, "y")
compare(x, y)

x <- "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Duis cursus
 tincidunt auctor. Vestibulum ac metus bibendum, facilisis nisi non, pulvinar
 dolor. Donec pretium iaculis nulla, ut interdum sapien ultricies a. "
y <- "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Duis cursus
 tincidunt auctor. Vestibulum ac metus1 bibendum, facilisis nisi non, pulvinar
 dolor. Donec pretium iaculis nulla, ut interdum sapien ultricies a. "
compare(x, y)
}

