\name{compare_state}
\alias{compare_state}
\title{Compare two directory states.}
\usage{compare_state(old, new)}

\description{
  Compare two directory states.
}
\value{list containing number of changes and files which have been 
\code{added}, \code{deleted} and \code{modified}}
\keyword{internal}
\arguments{
  \item{old}{previous state}
  \item{new}{current state}
}
