% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_name.R
\name{context_name}
\alias{context_name}
\title{Simplify a test filename to a context name.}
\usage{
context_name(filename)
}
\arguments{
\item{filename}{A character string giving the filename of a test file.
Standard prefix (\code{test-}) and suffix (\code{*.R}) are stripped.}
}
\description{
Simplify a test filename to a context name.
}
\examples{
context_name("test-my-function.R")
context_name("test-myFunction.R")
}
\keyword{internal}
