% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{context_start_file}
\alias{context_start_file}
\title{Start test context from a file name}
\usage{
context_start_file(name)
}
\arguments{
\item{name}{file name}
}
\description{
For use in external reporters
}
\keyword{internal}
