\name{error_report}
\alias{error_report}
\title{Generate error report from traceback.}
\usage{error_report(error, traceback)}

\description{
  Generate error report from traceback.
}
\keyword{internal}
\arguments{
  \item{error}{error message}
  \item{traceback}{traceback generated by \code{\link{create_traceback}}}
}
