% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectation.R
\name{exp_signal}
\alias{exp_signal}
\title{Signal an expectation}
\usage{
exp_signal(exp)
}
\arguments{
\item{exp}{An expectation object, as created by
\code{\link[=new_expectation]{new_expectation()}}.}
}
\description{
Signal an expectation
}
