% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock-object.R
\name{expect_args}
\alias{expect_args}
\title{Expectation: does the given argument values match expected?}
\usage{
expect_args(mock_object, n, ...)
}
\arguments{
\item{mock_object}{A \code{\link{mock}} object.}

\item{n}{Call number.}

\item{...}{Arguments as passed in a call.}
}
\description{
Unlike \code{\link{expect_call}}, \code{expect_args} compares the
actual values passed in to the call invoked on a \code{\link{mock}}
object.
}
\examples{
m <- mock()
a <- iris
with_mock(summary = m, summary(object = a))
expect_args(m, 1, object = a)
# is an equivalent to ...
expect_equal(m$args[[1]], list(object = a))
}
\seealso{
Other expectations: \code{\link{comparison-expectations}},
  \code{\link{equality-expectations}},
  \code{\link{expect_call}},
  \code{\link{expect_equal_to_reference}},
  \code{\link{expect_length}}, \code{\link{expect_match}},
  \code{\link{expect_named}},
  \code{\link{inheritance-expectations}},
  \code{\link{logical-expectations}},
  \code{\link{output-expectations}}
}

