% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-compiled-code.R
\name{expect_cpp_tests_pass}
\alias{expect_cpp_tests_pass}
\title{Test Compiled Code in a Package}
\usage{
expect_cpp_tests_pass(package)
}
\arguments{
\item{package}{The name of the package to test.}
}
\description{
Test compiled code in the package \code{package}. See
\code{\link[=use_catch]{use_catch()}} for more details.
}
\note{
A call to this function will automatically be generated
for you in \code{tests/testthat/test-cpp.R} after calling
\code{use_catch()}; you should not need to manually call
this expectation yourself.
}
