% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectations.r
\name{expect_is}
\alias{expect_is}
\alias{expect_type}
\title{Expectation: does the object inherit from a S3 or S4 class, or a base type?}
\usage{
expect_is(object, class, info = NULL, label = NULL)

expect_type(object, type)
}
\arguments{
\item{object}{object to test}

\item{class}{character vector of class names}

\item{info}{extra information to be included in the message (useful when
writing tests in loops).}

\item{label}{object label. When \code{NULL}, computed from deparsed object.}

\item{type}{String giving base type (as returned by \code{\link{typeof}}).}
}
\description{
Tests whether or not an object inherits from any of a list of classes, or
is an instance of a base type.
}
\examples{
expect_is(1, "numeric")
a <- matrix(1:10, nrow = 5)
expect_is(a, "matrix")

expect_is(mtcars, "data.frame")
# alternatively for classes that have an is method
expect_true(is.data.frame(mtcars))
}
\seealso{
\code{\link{inherits}}

Other expectations: \code{\link{equivalence}},
  \code{\link{expect-compare}},
  \code{\link{expect_equal_to_reference}},
  \code{\link{expect_named}}, \code{\link{expect_null}},
  \code{\link{expect_silent}},
  \code{\link{expect_takes_less_than}},
  \code{\link{expect_true}},
  \code{\link{matching-expectations}}
}

