% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect-comparison.R
\name{expect_less_than}
\alias{expect_less_than}
\alias{expect_more_than}
\title{Deprecated numeric comparison functions}
\usage{
expect_less_than(...)

expect_more_than(...)
}
\arguments{
\item{...}{All arguments passed on to \code{expect_lt()}/\code{expect_gt()}.}
}
\description{
These functions have been deprecated in favour of the more concise
\code{\link[=expect_gt]{expect_gt()}} and \code{\link[=expect_lt]{expect_lt()}}.
}
\keyword{internal}
