% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/expectations.r
\name{expect_message}
\alias{expect_message}
\title{Expectation: does expression show a message?}
\usage{
expect_message(object, regexp = NULL, ..., all = FALSE, info = NULL,
  label = NULL)
}
\arguments{
\item{object}{object to test}

\item{regexp}{optional regular expression to match. If not specified, just
asserts that expression shows some message.}

\item{...}{other arguments passed to \code{\link{expect_match}}}

\item{all}{if \code{TRUE}, all messages must match given regular expression;
if \code{FALSE} (the default), then only one message needs to match.}

\item{info}{extra information to be included in the message (useful when
writing tests in loops).}

\item{label}{object label. When \code{NULL}, computed from deparsed object.}
}
\description{
Needs to match at least one of the messages produced by the expression.
}
\examples{
f <- function(x) {
  if (x < 0) message("*x* is already negative")
  -x
}
expect_message(f(-1))
expect_message(f(-1), "already negative")
\dontrun{expect_message(f(1))}

# You can use the arguments of grepl to control the matching
expect_message(f(-1), "*x*", fixed = TRUE)
expect_message(f(-1), "NEGATIVE", ignore.case = TRUE)
}
\seealso{
Other expectations: \code{\link{equivalence}},
  \code{\link{expect_equal}},
  \code{\link{expect_equivalent}},
  \code{\link{expect_identical}};
  \code{\link{expect-compare}},
  \code{\link{expect_less_than}},
  \code{\link{expect_more_than}};
  \code{\link{expect_equal_to_reference}};
  \code{\link{expect_error}}; \code{\link{expect_false}},
  \code{\link{expect_true}}; \code{\link{expect_is}};
  \code{\link{expect_match}}; \code{\link{expect_named}};
  \code{\link{expect_null}}; \code{\link{expect_output}};
  \code{\link{expect_warning}};
  \code{\link{takes_less_than}}
}

