% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect-named.R
\name{expect_named}
\alias{expect_named}
\title{Do you expect a vector with (these) names?}
\usage{
expect_named(
  object,
  expected,
  ignore.order = FALSE,
  ignore.case = FALSE,
  info = NULL,
  label = NULL
)
}
\arguments{
\item{object}{Object to test.

Supports limited unquoting to make it easier to generate readable failures
within a function or for loop. See \link{quasi_label} for more details.}

\item{expected}{Character vector of expected names. Leave missing to
match any names. Use \code{NULL} to check for absence of names.}

\item{ignore.order}{If \code{TRUE}, sorts names before comparing to
ignore the effect of order.}

\item{ignore.case}{If \code{TRUE}, lowercases all names to ignore the
effect of case.}

\item{info}{Extra information to be included in the message. This argument
is soft-deprecated and should not be used in new code. Instead see
alternatives in \link{quasi_label}.}

\item{label}{Used to customise failure messages. For expert use only.}
}
\description{
You can either check for the presence of names (leaving \code{expected}
blank), specific names (by supplying a vector of names), or absence of
names (with \code{NULL}).
}
\examples{
x <- c(a = 1, b = 2, c = 3)
expect_named(x)
expect_named(x, c("a", "b", "c"))

# Use options to control sensitivity
expect_named(x, c("B", "C", "A"), ignore.order = TRUE, ignore.case = TRUE)

# Can also check for the absence of names with NULL
z <- 1:4
expect_named(z, NULL)
}
\seealso{
Other expectations:
\code{\link{comparison-expectations}},
\code{\link{equality-expectations}},
\code{\link[=expect_error]{expect_error()}},
\code{\link[=expect_length]{expect_length()}},
\code{\link[=expect_match]{expect_match()}},
\code{\link[=expect_null]{expect_null()}},
\code{\link[=expect_output]{expect_output()}},
\code{\link[=expect_reference]{expect_reference()}},
\code{\link[=expect_silent]{expect_silent()}},
\code{\link{inheritance-expectations}},
\code{\link{logical-expectations}}
}
\concept{expectations}
