% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect-self-test.R
\name{expect_no_success}
\alias{expect_no_success}
\alias{expect_no_failure}
\title{Test for absence of success or failure}
\usage{
expect_no_success(expr)

expect_no_failure(expr)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

These functions are deprecated because \code{\link[=expect_success]{expect_success()}} and
\code{\link[=expect_failure]{expect_failure()}} now test for exactly one success or no failures, and
exactly one failure and no successes.
}
\keyword{internal}
