% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress-output.R
\name{expect_output_file}
\alias{expect_output_file}
\title{Regression test: does output remain unchanged?}
\usage{
expect_output_file(object, file, update = TRUE, ..., info = NULL,
  label = NULL, print = FALSE, width = 80)
}
\arguments{
\item{object}{object to test}

\item{file}{Path to a "golden" text file that contains the desired output.}

\item{update}{Should the "golden" text file be updated? Defaults to
\code{FALSE}, but that it will automatically create output if the file
does not exist (i.e. on the first run).}

\item{...}{other values passed to \code{\link[=all.equal]{all.equal()}}}

\item{info}{extra information to be included in the message (useful when
writing tests in loops).}

\item{label}{object label. When \code{NULL}, computed from deparsed object.}

\item{print}{If \code{TRUE} and the result of evaluating \code{code} is
visible this will print the result, ensuring that the output of printing
the object is included in the overall output}

\item{width}{Number of characters per line of output}
}
\description{
It is often difficult to define exactly what the printed output should look
like, but you know it when you see it, and you want to make sure that it
doesn't change by accident. \code{expect_output_file()} renders function output
to a file then throws an error if the output changes.
}
\details{
You need to use \code{expect_output_file()} in conjunction with git or other
source code control tool so that you can see exactly what has changed, and
revert output to previous versions where needed.
}
\examples{
tmp <- tempfile()

# The first run always succeeds
expect_output_file(mtcars[1:10, ], tmp, print = TRUE)

# Subsequent runs will suceed only if the file is unchanged
# This will succeed:
expect_output_file(mtcars[1:10, ], tmp, print = TRUE)

\dontrun{
# This will fail
expect_output_file(mtcars[1:9, ], tmp, print = TRUE)
}
}
\seealso{
Other regression tests: \code{\link{expect_equal_to_reference}}
}
