% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect-length.R
\name{expect_shape}
\alias{expect_shape}
\title{Does code return an object with the specified shape?}
\usage{
expect_shape(object, shape, nrow, ncol)
}
\arguments{
\item{object}{Object to test.

  Supports limited unquoting to make it easier to generate readable failures
  within a function or for loop. See [quasi_label] for more details.}

\item{shape}{Expected shape, a numeric vector.}
\item{nrow}{Expected number of rows, numeric.}
\item{ncol}{Expected number of columns, numeric.}
}
\description{
By "shape", we mean an object's [dim()], or, for one-dimensional objects,
  it's [length()]. Thus this is an extension of [expect_length()] to more
  general objects like [data.frame()], [matrix()], and [array()].
To wit, first, the object's `dim()` is checked. If non-`NULL`, it is compared
  to `shape` (or one/both of `nrow`, `ncol`, if they are supplied, in which
  case they take precedence). If `dim(object)` is `NULL`, `length(object)`
  is compared to `shape`.
}
\seealso{
[expect_length()] to specifically make assertions about the
  [length()] of a vector.

Other expectations: 
\code{\link{comparison-expectations}},
\code{\link{equality-expectations}},
\code{\link{expect_error}()},
\code{\link{expect_length}()},
\code{\link{expect_match}()},
\code{\link{expect_named}()},
\code{\link{expect_null}()},
\code{\link{expect_output}()},
\code{\link{expect_reference}()},
\code{\link{expect_silent}()},
\code{\link{inheritance-expectations}},
\code{\link{logical-expectations}}
}
\concept{expectations}
