% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot-file.R
\name{expect_snapshot_file}
\alias{expect_snapshot_file}
\title{Snapshot testing for a complete file}
\usage{
expect_snapshot_file(path, name = basename(path), binary = TRUE, cran = FALSE)
}
\arguments{
\item{path}{Path to file to snapshot.}

\item{name}{Snapshot name, taken from \code{path} by default.}

\item{binary}{If \code{FALSE}, files are compared line-by-line, ignoring the
difference between Windows and Mac/Linux line endings.}

\item{cran}{Should these expectations be verified on CRAN? By default,
they are not, because snapshot tests tend to be fragile because they
often rely on minor details of dependencies. Even on CRAN, the
expectations will still fail if \code{x} errors.}
}
\description{
Whole file snapshot testing is primarily designed for testing images.
The first time it is run \code{expect_snapshot_file()} will create
\verb{_snaps/\{test\}/\{name\}.\{ext\}} containing reference output. Future runs will be
compared to this reference: if different, the test will fail and the new
results will be saved in \verb{_snaps/\{test\}/\{name\}.new.\{ext\}}.

We generally expect this function to be used via a wrapper that takes care
of ensuring the images are as reproducible as possible, automatically
skipping tests where it's known that the images can't be reproduced exactly
(without extreme efforts it is generally impossible to generate pixel
perfect images across platforms).
}
