% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectations.r
\name{expect_takes_less_than}
\alias{expect_takes_less_than}
\title{Expectation: does expression take less than a fixed amount of time to run?}
\usage{
expect_takes_less_than(object, amount, info = NULL, label = NULL)
}
\arguments{
\item{object}{object to test}

\item{amount}{maximum duration in seconds}

\item{info}{extra information to be included in the message (useful when
writing tests in loops).}

\item{label}{object label. When \code{NULL}, computed from deparsed object.}
}
\description{
This is useful for performance regression testing.
}
\seealso{
Other expectations: \code{\link{equivalence}},
  \code{\link{expect-compare}},
  \code{\link{expect_equal_to_reference}},
  \code{\link{expect_is}}, \code{\link{expect_named}},
  \code{\link{expect_null}}, \code{\link{expect_silent}},
  \code{\link{expect_true}},
  \code{\link{matching-expectations}}
}

