% Generated by roxygen2 (4.0.2): do not edit by hand
\name{expect_that_combine}
\alias{expect_that_combine}
\title{alternative \code{expect_that} from \code{testthat} which permutes all
  the inputs to a function which should give the same result where n args >=2 and the function is commutative.}
\usage{
expect_that_combine(object, condition, info = NULL, label = NULL)
}
\value{
testthat result
}
\description{
This makes a lot of assumptions, needs more testing. It can't
  handle mixed error/no error outcomes after permutation, which is an
  important feature to consider. The command following this function attaches this function to the testthat namespace. This means that it can call internal testthat functions, but does not mean it appears as testthat::expect_that_combine
}
\examples{
expect_that_combine(sum(1,2,3), testthat::equals(6))
}

