% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectation.r
\name{expectation}
\alias{expectation}
\alias{is.expectation}
\title{Expectation class.}
\usage{
expectation(passed, message, srcref = NULL)

is.expectation(x)
}
\arguments{
\item{passed}{a single logical value indicating whether the test passed
(\code{TRUE}), failed (\code{FALSE}), or threw an error (\code{NA})}

\item{message}{A text description of failure}

\item{srcref}{Source reference, if known}

\item{x}{object to test for class membership}
}
\description{
Any expectation should return objects of this class - see the built in
expectations for details.
}
\keyword{internal}

