% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-fail.R
\docType{class}
\name{FailReporter-class}
\alias{FailReporter}
\alias{FailReporter-class}
\title{Test reporter: fail at end.}
\arguments{
\item{...}{Arguments used to initialise class}
}
\description{
This reporter will simply throw an error if any of the tests failed. It is
best combined with another reporter, such as the
\code{\link{SummaryReporter}}.
}
\keyword{debugging}

