% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-files.R
\name{filter_test_scripts}
\alias{filter_test_scripts}
\title{Filter File List for Tests}
\usage{
filter_test_scripts(files, filter = NULL, invert = FALSE, ...)
}
\arguments{
\item{files}{character vector of file names to filter}

\item{filter}{cf \code{\link{test_dir}}}

\item{invert}{If \sQuote{TRUE} return files which do \emph{not} match.}

\item{...}{Additional arguments passed to \code{grepl} to control filtering.}
}
\value{
the test file paths
}
\description{
Filtering is done by \code{grep}ing with pattern \code{filter} against
the basename from \code{files} with any leading \sQuote{test-?} and trailing
\sQuote{.[Rr]} removed.
}
\keyword{internal}

