% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-zzz.r
\name{find_reporter}
\alias{find_reporter}
\title{Find reporter object given name or object.}
\usage{
find_reporter(reporter)
}
\arguments{
\item{reporter}{name of reporter(s)}
}
\description{
If not found, will return informative error message.
This function supports the aggregation of reporters into a
\code{\link{MultiReporter}}: pass a character vector or separate the
individual reporters by \code{"+"}.
Will return null if given NULL.
}
\keyword{internal}

