\name{find_test_scripts}
\alias{find_test_scripts}
\title{take care or finding the test files and sourcing the helpers}
\usage{
  find_test_scripts(path, filter = NULL)
}
\arguments{
  \item{path}{path to tests}

  \item{filter}{cf \code{\link{test_dir}}}
}
\value{
  the test file paths
}
\description{
  take care or finding the test files and sourcing the
  helpers
}

