% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{is_snapshotting}
\alias{is_snapshotting}
\title{Is the code currently run within a snapshot test?}
\usage{
is_snapshotting()
}
\description{
This is often useful to compute a default argument for \code{quiet}. You
often want to deliberatey suppress output for tests but still track
it when snapshotting: \code{quiet = is_testing() && !is_snapshotting() }
}
\keyword{internal}
