% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-env.R
\name{is_testing}
\alias{is_testing}
\alias{is_parallel}
\alias{is_snapshot}
\alias{testing_package}
\title{Determine testing status}
\usage{
is_testing()

is_parallel()

is_snapshot()

testing_package()
}
\description{
These functions help you determine if you code is running in a particular
testing content:
\itemize{
\item \code{is_testing()} is \code{TRUE} when run inside a test.
\item \code{is_parallel()} is \code{TRUE} if the test is run in parallel.
\item \code{is_snapshot()} is \code{TRUE} when run inside a snapshot test.
\item \code{testing_package()} gives name of the package being tested.
}

A common use of these functions is to compute a default value for a \code{quiet}
argument with \code{is_testing() && !is_snapshotting()}. In this case, you'll
want to avoid an run-time dependency on testthat, in which case you should
just copy the implementation of these functions in a \code{utils.R} or similar.
}
