% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/library.r
\name{library_if_available}
\alias{library_if_available}
\title{Load package, if available.}
\usage{
library_if_available(package)
}
\arguments{
\item{package}{package name (without quotes)}
}
\description{
Quietly load a package if it is installed, otherwise do nothing.  This is
useful for testing files so that you can run them while you are developing
your package, before it is installed for the first time; then continue to
have the same code work when the tests are run automatically by R CMD
CHECK.
}
\examples{
\dontrun{
library_if_available(testthat)
library_if_available(packagethatdoesntexist)
}
}

