% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-llm.R
\name{LlmReporter}
\alias{LlmReporter}
\title{Report test progress for LLMs}
\description{
\code{LlmReporter} is designed for use with Large Language Models (LLMs).
It reports problems (warnings, skips, errors, and failures) as they
occur and the total number of successes at the end.

\code{LlmReporter} is used by default when tests are run by a coding agent.
Currently we detect Posit Assistant (via \code{getOption("posit_assistant")}),
Claude Code, Codex, Cursor, and Gemini CLI (via env vars).
If using another tool, configure it to set env var \code{AGENT=1}.
}
\seealso{
Other reporters:
\code{\link{CheckReporter}},
\code{\link{DebugReporter}},
\code{\link{FailReporter}},
\code{\link{JunitReporter}},
\code{\link{ListReporter}},
\code{\link{LocationReporter}},
\code{\link{MinimalReporter}},
\code{\link{MultiReporter}},
\code{\link{ProgressReporter}},
\code{\link{RStudioReporter}},
\code{\link{Reporter}},
\code{\link{SilentReporter}},
\code{\link{SlowReporter}},
\code{\link{StopReporter}},
\code{\link{SummaryReporter}},
\code{\link{TapReporter}},
\code{\link{TeamcityReporter}}
}
\concept{reporters}
