% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edition.R
\name{local_edition}
\alias{local_edition}
\title{Temporarily change the active testthat edition}
\usage{
local_edition(x, .env = parent.frame())
}
\arguments{
\item{x}{Edition Should be a single integer.}

\item{.env}{Environment that controls scope of changes. For expert use only.}
}
\description{
\code{local_edition()} allows you to temporarily (within a single test or
a single test file) change the active edition of testthat.
}
\keyword{internal}
