% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.R
\name{local_test_directory}
\alias{local_test_directory}
\title{Locally set test directory options}
\usage{
local_test_directory(path, package = NULL, .env = parent.frame())
}
\arguments{
\item{path}{Path to directory of files}

\item{package}{Optional package name, if known.}
}
\description{
For expert use only.
}
\keyword{internal}
