% Generated by roxygen2 (4.0.2): do not edit by hand
\name{make_expectation}
\alias{make_expectation}
\title{Make an equality test.}
\usage{
make_expectation(x, expectation = "equals")
}
\arguments{
\item{x}{a vector of values}

\item{expectation}{the type of equality you want to test for
(\code{equals}, \code{is_equivalent_to}, \code{is_identical_to})}
}
\description{
This a convenience function to make a expectation that checks that
input stays the same.
}
\examples{
x <- 1:10
make_expectation(x)

make_expectation(mtcars$mpg)

df <- data.frame(x = 2)
make_expectation(df)
}

