% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reporter-minimal.r
\docType{class}
\name{MinimalReporter-class}
\alias{MinimalReporter}
\alias{MinimalReporter-class}
\title{Test reporter: minimal.}
\arguments{
\item{...}{Arguments used to initialise class}
}
\description{
The minimal test reporter provides the absolutely minimum amount of
information: whether each expectation has succeeded, failed or experienced
an error.  If you want to find out what the failures and errors actually
were, you'll need to run a more informative test reporter.
}
\keyword{debugging}

