\docType{class}
\name{MinimalReporter-ref-class}
\alias{MinimalReporter-class}
\alias{MinimalReporter-ref-class}
\title{Test reporter: minimal.}
\arguments{
  \item{...}{Arguments used to initialise class}
}
\description{
The minimal test reporter provides the absolutely minimum
amount of information: whether each expectation has
succeeded, failed or experienced an error.  If you want to
find out what the failures and errors actually were, you'll
need to run a more informative test reporter.
}
\keyword{debugging}

