\name{MinimalSuite}
\alias{MinimalSuite}
\title{Test reporter: minimal...}

\description{
  Test reporter: minimal
}

\details{
  The minimal test suite provides the absolutely minimum amount of
  information: whether each expectation has succeeded, failed or
  experienced an error.  If you want to find out what the failures and
  errors actually were, you'll need to run a more informative test
  suite.
}
\keyword{debugging}
