% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-otel.R
\name{OpenTelemetryReporter}
\alias{OpenTelemetryReporter}
\title{OpenTelemetry reporter: traces for test results}
\arguments{
\item{pkg}{A path to an R package, by default the one in the current
directory.}

\item{tracer}{An \pkg{otel} tracer, or \code{NULL} to use the default tracer.}
}
\description{
A variant of the Check reporter that also emits OpenTelemetry traces for
tests. Span attributes are drawn from \href{https://opentelemetry.io/docs/specs/semconv/registry/attributes/test/}{the semantic conventions for tests}.
}
\seealso{
Other reporters: 
\code{\link{CheckReporter}},
\code{\link{DebugReporter}},
\code{\link{FailReporter}},
\code{\link{JunitReporter}},
\code{\link{ListReporter}},
\code{\link{LocationReporter}},
\code{\link{MinimalReporter}},
\code{\link{MultiReporter}},
\code{\link{ProgressReporter}},
\code{\link{RStudioReporter}},
\code{\link{Reporter}},
\code{\link{SilentReporter}},
\code{\link{StopReporter}},
\code{\link{SummaryReporter}},
\code{\link{TapReporter}},
\code{\link{TeamcityReporter}}
}
\concept{reporters}
