% Generated by roxygen2 (4.0.0): do not edit by hand
\name{pending_expectation}
\alias{format.pending_expectation}
\alias{is.pending_expectation}
\alias{pending_expectation}
\alias{print.pending_expectation}
\title{Pending_Expectation class.}
\usage{
pending_expectation(pending_msg = "unknown")

is.pending_expectation(x)
}
\arguments{
  \item{pending_msg}{A text description of pending}

  \item{x}{object to test for class membership}
}
\description{
A pending expectation signals that a test has a pending implementation
and is currently inconclusive.
The pending expectation is also a successful \code{\link{expectation}}.
}
\keyword{internal}

