% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.testthat_results}
\alias{print.testthat_results}
\title{print testthat_results as a data frame}
\usage{
\method{print}{testthat_results}(x, ...)
}
\arguments{
\item{x}{the testthat results}

\item{...}{ignored}
}
\description{
print testthat_results as a data frame
}

