% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-directory-c.R
\name{process_setup}
\alias{process_setup}
\title{This function is called from the subprocess at startup}
\usage{
process_setup(path, package, testthat_dir, load_helpers, pkgload = FALSE)
}
\arguments{
\item{package}{Name of the tested package. This is needed, because we
need to attach the package in the subprocess.}

\item{testthat_dir}{The root directory of the package being tested.}

\item{pkgload}{Whether to use \code{pkgload::load_all()} or \code{library()} to
attach the tested package.}
}
\description{
This function is called from the subprocess at startup
}
\keyword{internal}
