% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reporter-rstudio.R
\docType{class}
\name{RstudioReporter-class}
\alias{RstudioReporter}
\alias{RstudioReporter-class}
\title{Test reporter: RStudio}
\arguments{
\item{...}{Arguments used to initialise class}
}
\description{
This reporter is designed for output to RStudio. It produces results in
any easily parsed form.
}
\keyword{debugging}

