% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-compiled-code.R
\name{expect_cpp_tests_pass}
\alias{expect_cpp_tests_pass}
\alias{run_cpp_tests}
\title{Expectation: do C++ tests past?}
\usage{
expect_cpp_tests_pass(package)

run_cpp_tests(package)
}
\arguments{
\item{package}{The name of the package to test.}
}
\description{
Test compiled code in the package \code{package}. A call to this function will
automatically be generated for you in \code{tests/testthat/test-cpp.R} after
calling \code{\link[=use_catch]{use_catch()}}; you should not need to manually call this expectation
yourself.
}
\keyword{internal}
