% Generated by roxygen2 (4.0.2): do not edit by hand
\name{safe_digest}
\alias{safe_digest}
\title{Compute a digest of a filename, returning NA if the file doesn't
exist.}
\usage{
safe_digest(path)
}
\arguments{
\item{filename}{filename to compute digest on}
}
\value{
a digest of the file, or NA if it doesn't exist.
}
\description{
Compute a digest of a filename, returning NA if the file doesn't
exist.
}
\keyword{internal}

