% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-progress.R
\name{set_max_fails}
\alias{set_max_fails}
\title{Set maximum number of test failures allowed before aborting the run}
\usage{
set_max_fails(n)
}
\arguments{
\item{n}{Maximum number of failures allowed.}
}
\description{
This sets the \code{TESTTHAT_MAX_FAILS} env var which will affect both the
current R process and any processes launched from it.
}
\keyword{internal}
