% Generated by roxygen2 (4.0.2): do not edit by hand
\name{skip}
\alias{skip}
\alias{skip_on_cran}
\title{Skip a test.}
\usage{
skip(message)

skip_on_cran()
}
\arguments{
\item{message}{An message describing why the test was skipped.}
}
\description{
This function allows you to skip a test if it's not currently available.
This will produce an informative message, but will not cause the test
suite to fail. \code{skip_on_cran()} uses \code{skip()}, together with
the \code{NOT_CRAN} environment variable (set by devtools), to automatically
skip tests that should not be run on CRAN.
}
\examples{
if (FALSE) skip("No internet connection")
}

