% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skip.R
\name{skip_on_travis}
\alias{skip_on_travis}
\alias{skip_on_appveyor}
\title{Superseded skip functions}
\usage{
skip_on_travis()

skip_on_appveyor()
}
\description{
\itemize{
\item \code{skip_on_travis()} and \code{skip_on_appveyor()} have been superseded by
\code{skip_on_ci()}.
}
}
\keyword{internal}
