% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{snapshot_accept}
\alias{snapshot_accept}
\title{Snapshot management}
\usage{
snapshot_accept(path = "tests/testthat")
}
\arguments{
\item{path}{Path to tests}
}
\description{
\code{snapshot_accept()} accepts all changed snapshots.
}
