% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-sonarqube.R
\name{SonarqubeReporter}
\alias{SonarqubeReporter}
\title{Report results in SonarQube generic test execution XML format}
\description{
This reporter produces XML output following the SonarQube
\href{https://docs.sonarsource.com/sonarqube-server/analyzing-source-code/test-coverage/generic-test-data#generic-test-execution}{Generic Test Execution Report Format},
written to a file (or stdout). The resulting XML can be imported into
SonarQube by setting the \code{sonar.testExecutionReportPaths} analysis
parameter. Requires the \emph{xml2} package.

Test files become \verb{<file>} elements (with the path attribute set to the
relative test file path), and individual \code{test_that()} blocks become
\verb{<testCase>} elements. On failure, error, or skip, a corresponding
child element (\verb{<failure>}, \verb{<error>}, or \verb{<skipped>}) is added to the
test case.
}
\seealso{
Other reporters: 
\code{\link{CheckReporter}},
\code{\link{DebugReporter}},
\code{\link{FailReporter}},
\code{\link{JunitReporter}},
\code{\link{ListReporter}},
\code{\link{LlmReporter}},
\code{\link{LocationReporter}},
\code{\link{MinimalReporter}},
\code{\link{MultiReporter}},
\code{\link{ProgressReporter}},
\code{\link{RStudioReporter}},
\code{\link{Reporter}},
\code{\link{SilentReporter}},
\code{\link{SlowReporter}},
\code{\link{StopReporter}},
\code{\link{SummaryReporter}},
\code{\link{TapReporter}},
\code{\link{TeamcityReporter}}
}
\concept{reporters}
