% Generated by roxygen2 (4.0.2): do not edit by hand
\name{source_test_helpers}
\alias{source_test_helpers}
\title{source the helper scripts if any}
\usage{
source_test_helpers(path, env = globalenv())
}
\arguments{
\item{path}{path to tests}

\item{env}{environment in which to execute test suite.}
}
\description{
Helper scripts are R scripts accompanying test scripts
but prefixed by \code{helper}. These scripts are sourced
only one time in test environment.
}

