% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect-that.R
\name{succeed}
\alias{succeed}
\title{Mark a test as successful}
\usage{
succeed(message = "Success has been forced", info = NULL)
}
\arguments{
\item{message}{A character vector describing the failure. The
first element should describe the expected value, and the second (and
optionally subsequence) elements should describe what was actually seen.}

\item{info}{Character vector continuing additional information. Included
for backward compatibility only and new expectations should not use it.}
}
\description{
This is an older version of \code{\link[=pass]{pass()}} that exists for backwards compatibility.
You should now use \code{pass()} instead.
}
\keyword{internal}
