% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{SummaryReporter-class}
\alias{SummaryReporter}
\alias{SummaryReporter-class}
\title{Test reporter: summary of errors.}
\arguments{
\item{...}{Arguments used to initialise class}
}
\description{
This is the most useful reporting reporter as it lets you know both which
tests have run successfully, as well as fully reporting information about
failures and errors.  It is the default reporting reporter used by
\code{\link{test_dir}} and \code{\link{test_file}}.
}
\details{
You can use the \code{max_reports} field to control the maximum number
of detailed reports produced by this reporter. This is useful when running
with \code{\link{auto_test}}

As an additional benefit, this reporter will praise you from time-to-time
if all your tests pass.
}
\keyword{debugging}

