% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reporter-tap.r
\docType{class}
\name{TapReporter-class}
\alias{TapReporter}
\alias{TapReporter-class}
\title{Test reporter: TAP format.}
\arguments{
\item{...}{Arguments used to initialise class}
}
\description{
This reporter will output results in the Test Anything Protocol (TAP),
a simple text-based interface between testing modules in a test harness.
For more information about TAP, see http://testanything.org
}
\keyword{debugging}

