\docType{class}
\name{TapReporter-ref-class}
\alias{TapReporter-class}
\alias{TapReporter-ref-class}
\title{Test reporter: TAP format.}
\arguments{
  \item{...}{Arguments used to initialise class}
}
\description{
This reporter will output results in the Test Anything
Protocol (TAP), a simple text-based interface between
testing modules in a test harness. For more information
about TAP, see http://testanything.org
}
\keyword{debugging}

