% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teardown.R
\name{teardown}
\alias{teardown}
\title{Run code on teardown}
\usage{
teardown(code, env = parent.frame())
}
\arguments{
\item{code}{Code to evaluate}

\item{env}{Environment in which code will be evaluted. For expert
use only.}
}
\description{
Code in a teardown block is guaranteed to be run upon completion of a test
file. Multiple calls to \code{teardown} will be executed in the order they
were created.
}
\examples{
\dontrun{
teardown{
  tmp <- tempfile()
  writeLines(tmp, "some test data")
  teardown(unlink(tmp))
}
}
}
