% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test-compiled-code.r
\name{test_compiled_code}
\alias{test_compiled_code}
\title{Test Compiled Code in a Package}
\usage{
test_compiled_code(package)
}
\arguments{
\item{package}{The name of the package to test.}
}
\description{
Test compiled code in the package \code{package}. See
\code{\link{use_catch}()} for more details.
}

