% Generated by roxygen2 (4.0.2): do not edit by hand
\name{test_coverage}
\alias{test_coverage}
\title{Run all tests in an installed package to check test code coverage}
\usage{
test_coverage(package, filter = NULL)
}
\arguments{
\item{package}{package name}

\item{filter}{If not \code{NULL}, only tests with file names matching this
regular expression will be executed.  Matching will take on the file
name after it has been stripped of \code{"test-"} and \code{".r"}.}
}
\value{
a data frame of the summary of test results
}
\description{
Run \code{test_package} silently, just looking for function calls, then
determine what proportion of functions in the tested package were tested.
}
\examples{
\dontrun{test_package_cover("testthat")}
}

