% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat-package.R
\docType{package}
\name{testthat-package}
\alias{testthat}
\alias{testthat-package}
\title{An R package to make testing fun!}
\description{
Try the example below. Have a look at the references and learn more
from function documentation such as \code{\link[=test_that]{test_that()}}.
}
\section{Options}{

\itemize{
\item \code{testthat.use_colours}: Should the output be coloured? (Default: \code{TRUE}).
\item \code{testthat.summary.max_reports}: The maximum number of detailed test
reports printed for the summary reporter (default: 10).
\item \code{testthat.summary.omit_dots}: Omit progress dots in the summary reporter
(default: \code{FALSE}).
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://testthat.r-lib.org}
  \item \url{https://github.com/r-lib/testthat}
  \item Report bugs at \url{https://github.com/r-lib/testthat/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@posit.co}

Authors:
\itemize{
  \item Hadley Wickham \email{hadley@posit.co}
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
  \item R Core team (Implementation of utils::recover()) [contributor]
}

}
\keyword{internal}
