% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{testthat_examples}
\alias{testthat_examples}
\alias{testthat_example}
\title{Retrieve paths to built-in example test files}
\usage{
testthat_examples()

testthat_example(filename)
}
\arguments{
\item{filename}{Name of test file}
}
\description{
\code{testthat_examples()} retrieves path to directory of test files,
\code{testthat_example()} retrieves path to a single test file.
}
\examples{
dir(testthat_examples())
testthat_example("success")
}
\keyword{internal}
