% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{testthat_tolerance}
\alias{testthat_tolerance}
\title{Default numeric tolerance}
\usage{
testthat_tolerance()
}
\description{
testthat's default numeric tolerance is 1.4901161 × 10\if{html}{\out{<sup>}}-8\if{html}{\out{</sup>}}.
}
\keyword{internal}
