% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-timing.R
\name{TimingReporter}
\alias{TimingReporter}
\title{Test reporter: show timings for slow tests}
\description{
\code{TimingReporter} is designed to identify slow tests. It reports the
execution time for each test and can optionally filter out tests that
run faster than a specified threshold (default: 1 second). This reporter
is useful for performance optimization and identifying tests that may
benefit from optimization or parallelization.
}
\seealso{
Other reporters: 
\code{\link{CheckReporter}},
\code{\link{DebugReporter}},
\code{\link{FailReporter}},
\code{\link{JunitReporter}},
\code{\link{ListReporter}},
\code{\link{LocationReporter}},
\code{\link{MinimalReporter}},
\code{\link{MultiReporter}},
\code{\link{ProgressReporter}},
\code{\link{RStudioReporter}},
\code{\link{Reporter}},
\code{\link{SilentReporter}},
\code{\link{StopReporter}},
\code{\link{SummaryReporter}},
\code{\link{TapReporter}},
\code{\link{TeamcityReporter}}
}
\concept{reporters}
