% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{simulate_test_env}
\alias{simulate_test_env}
\title{Simulate a test environment}
\usage{
simulate_test_env(package, path)
}
\description{
This function is designed to allow you to simulate testthat's testing
environment in an interactive session. To undo it's affect, you
will need to restart your R session.
}
\keyword{internal}
