% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{VerboseReporter-class}
\alias{VerboseReporter}
\alias{VerboseReporter-class}
\title{Test reporter: verbose test progress}
\arguments{
\item{...}{Arguments used to initialise class}
}
\description{
This reporter lets you see detail of errors quickly.
}
\details{
You can use the \code{max_reports} field to control the maximum number
of detailed reports produced by this reporter. This is useful when running
with \code{\link{auto_test}}

As an additional benefit, this reporter will praise you from time-to-time
if all your tests pass.
}
\keyword{debugging}

