% Generated by roxygen2 (4.0.1): do not edit by hand
\name{with_mock}
\alias{with_mock}
\title{Mock functions in a package.}
\usage{
with_mock(.code, ..., .env = topenv())
}
\arguments{
\item{.code}{the code to be executed}

\item{...}{named parameters giving definitions of mocked functions}

\item{.env}{the environment in which to patch the functions,
defaults to the top-level environment}
}
\description{
Executes code after temporarily substituting implementations of package
functions.  This is useful for testing code that relies on functions that are
slow, have unintended side effects or access resources that may not be
available when testing.
}
\examples{
with_mock(
  expect_equal(3, 5),
  all.equal = function(x, y, ...) TRUE,
  .env = asNamespace("base"))
throws_error()(expect_equal(3, 5))
}

