% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.data.frame.testthat_results}
\alias{as.data.frame.testthat_results}
\title{convert testthat_results to data frame}
\usage{
\method{as.data.frame}{testthat_results}(x, ...)
}
\arguments{
\item{x}{the testthat results}

\item{...}{ignored}
}
\value{
a data frame
}
\description{
convert testthat_results to data frame
}

