% Generated by roxygen2 (4.0.2): do not edit by hand
\name{auto_test_package}
\alias{auto_test_package}
\title{Watches a package for changes, rerunning tests as appropriate.}
\usage{
auto_test_package(pkg = ".", reporter = "summary")
}
\arguments{
\item{pkg}{path to package}

\item{reporter}{test reporter to use}
}
\description{
Watches a package for changes, rerunning tests as appropriate.
}
\seealso{
\code{\link{auto_test}} for details on how method works
}
\keyword{debugging}

