% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reporter-check.R
\docType{class}
\name{CheckReporter-class}
\alias{CheckReporter}
\alias{CheckReporter-class}
\title{Check reporter: 13 line summary of problems}
\arguments{
\item{...}{Arguments used to initialise class}
}
\description{
\code{R CMD check} displays only the last 13 lines of the result, so this
report is design to ensure that you see something useful there.
}
\keyword{debugging}

