% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-check.R
\docType{data}
\name{CheckReporter}
\alias{CheckReporter}
\title{Check reporter: 13 line summary of problems}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CheckReporter
}
\description{
\code{R CMD check} displays only the last 13 lines of the result, so this
report is design to ensure that you see something useful there.
}
\keyword{datasets}

