% Generated by roxygen2 (4.0.1): do not edit by hand
\name{colourise}
\alias{colourise}
\title{Colourise text for display in the terminal.}
\usage{
colourise(text, fg = "black", bg = NULL)
}
\arguments{
\item{text}{character vector}

\item{fg}{foreground colour, defaults to white}

\item{bg}{background colour, defaults to transparent}
}
\description{
If R is not currently running in a system that supports terminal colours
the text will be returned unchanged.
}
\details{
Allowed colours are: black, blue, brown, cyan, dark gray, green, light
blue, light cyan, light gray, light green, light purple, light red,
purple, red, white, yellow
}
\examples{
print(colourise("Red", "red"))
cat(colourise("Red", "red"), "\\n")
cat(colourise("White on red", "white", "red"), "\\n")
}

