% Generated by roxygen2 (4.0.2): do not edit by hand
\name{compare_state}
\alias{compare_state}
\title{Compare two directory states.}
\usage{
compare_state(old, new)
}
\arguments{
\item{old}{previous state}

\item{new}{current state}
}
\value{
list containing number of changes and files which have been
  \code{added}, \code{deleted} and \code{modified}
}
\description{
Compare two directory states.
}
\keyword{internal}

