% Generated by roxygen2 (4.0.2): do not edit by hand
\name{context}
\alias{context}
\title{Describe the context of a set of tests.}
\usage{
context(desc)
}
\arguments{
\item{desc}{description of context.  Should start with a capital letter.}
}
\description{
A context defines a set of tests that test related functionality.  Usually
you will have one context per file, but you may have multiple contexts
in a single file if you so choose.
}
\examples{
context("String processing")
context("Remote procedure calls")
}

