% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{context}
\alias{context}
\title{Describe the context of a set of tests.}
\usage{
context(desc)
}
\arguments{
\item{desc}{description of context.  Should start with a capital letter.}
}
\description{
Use of \code{context()} is no longer recommended. Instead omit it, and messages
will use the name of the file instead. This ensures that the context and
test file name are always in sync.
}
\details{
A context defines a set of tests that test related functionality.  Usually
you will have one context per file, but you may have multiple contexts
in a single file if you so choose.
}
\examples{
context("String processing")
context("Remote procedure calls")
}
\keyword{internal}
