% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-debug.R
\docType{data}
\name{DebugReporter}
\alias{DebugReporter}
\title{Test reporter: start recovery.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DebugReporter
}
\description{
This reporter will call a modified version of \code{\link[=recover]{recover()}} on all
broken expectations.
}
\seealso{
Other reporters: \code{\link{CheckReporter}},
  \code{\link{FailReporter}}, \code{\link{ListReporter}},
  \code{\link{LocationReporter}},
  \code{\link{MinimalReporter}},
  \code{\link{MultiReporter}},
  \code{\link{ProgressReporter}},
  \code{\link{RstudioReporter}},
  \code{\link{SilentReporter}}, \code{\link{StopReporter}},
  \code{\link{SummaryReporter}}, \code{\link{TapReporter}},
  \code{\link{TeamcityReporter}}
}
\keyword{datasets}
