% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-debug.R
\docType{data}
\name{DebugReporter}
\alias{DebugReporter}
\title{Test reporter: start recovery.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DebugReporter
}
\description{
This reporter will call a modified version of \code{\link{recover}} on all
broken expectations.
}
\keyword{datasets}

