% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter.R
\name{default_reporter}
\alias{default_reporter}
\title{Retrieve the default reporter.}
\usage{
default_reporter()
}
\description{
This will be \code{\link[=SummaryReporter]{SummaryReporter()}} unless option
\code{testthat.default_reporter} is set
}
\keyword{internal}
