% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter.R
\name{default_reporter}
\alias{default_reporter}
\alias{default_compact_reporter}
\alias{check_reporter}
\title{Retrieve the default reporter}
\usage{
default_reporter()

default_compact_reporter()

check_reporter()
}
\description{
The defaults are:
\itemize{
\item \link{ProgressReporter} for interactive; override with \code{testthat.default_reporter}
\item \link{CompactProgressReporter} for single-file interactive; override with
\code{testthat.default_compact_reporter}
\item \link{CheckReporter} for R CMD check; override with \code{testthat.default_check_reporter}
}
}
\keyword{internal}
