% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dir_state}
\alias{dir_state}
\title{Capture the state of a directory.}
\usage{
dir_state(path, pattern = NULL, hash = TRUE)
}
\arguments{
\item{path}{path to directory}

\item{pattern}{regular expression with which to filter files}

\item{hash}{use hash (slow but accurate) or time stamp (fast but less
accurate)}
}
\description{
Capture the state of a directory.
}
\keyword{internal}

