% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect-equality.R
\name{equality-expectations}
\alias{equality-expectations}
\alias{expect_equal}
\alias{expect_equivalent}
\alias{expect_identical}
\title{Expectation: is the object equal to a value?}
\usage{
expect_equal(
  object,
  expected,
  ...,
  tolerance = testthat_tolerance(),
  info = NULL,
  label = NULL,
  expected.label = NULL
)

expect_equivalent(
  object,
  expected,
  ...,
  info = NULL,
  label = NULL,
  expected.label = NULL,
  tolerance = testthat_tolerance()
)

expect_identical(
  object,
  expected,
  info = NULL,
  label = NULL,
  expected.label = NULL,
  ...
)
}
\arguments{
\item{object, expected}{Computation and value to compare it to.

Both arguments supports limited unquoting to make it easier to generate
readable failures within a function or for loop. See \link{quasi_label} for
more details.}

\item{...}{Passed on to \code{\link[waldo:compare]{waldo::compare()}} to control the details of
the comparison.}

\item{tolerance}{If non-\code{NULL}, used as threshold for ignoring small
floating point difference when comparing numeric vectors. Setting to
any non-\code{NULL} value will cause integer and double vectors to be compared
based on their values, rather than their types.

It uses the same algorithm as \code{\link[=all.equal]{all.equal()}}, i.e., first we generate
\code{x_diff} and \code{y_diff} by subsetting \code{x} and \code{y} to look only locations
with differences. Then we check that
\code{mean(abs(x_diff - y_diff)) / mean(abs(y_diff))} (or just
\code{mean(abs(x_diff - y_diff))} if \code{y_diff} is small) is less than
\code{tolerance}.}

\item{info}{Extra information to be included in the message. This argument
is soft-deprecated and should not be used in new code. Instead see
alternatives in \link{quasi_label}.}

\item{label, expected.label}{Used to customise failure messages. For expert
use only.}
}
\description{
These three functions use \href{https://github.com/r-lib/waldo}{waldo} to
compare the results of some code to an \code{expected} result. They differ in
their level of strictness, as determined by additional arguments passed on
to \code{waldo::compare()}:
\itemize{
\item \code{expect_identical()} is the strictest comparison.
\item \code{expect_equal()} sets \code{tolerance = testthat_tolerance()} so that
small floating point differences are ignored; this also implies that
(e.g.) \code{1} and \code{1L} are treated as equal.
\item \code{expect_equivalent()} sets \code{ignore_attr = TRUE} and
\code{tolerance = testthat_tolerance()}, ignoring small floating point
differences and all differences in attributes.
}

In most cases you should use \code{expect_equal()}.
}
\examples{
a <- 10
expect_equal(a, 10)

# Use expect_equal() when testing for numeric equality
\dontrun{
expect_identical(sqrt(2) ^ 2, 2)
}
expect_equal(sqrt(2) ^ 2, 2)

# expect_equivalent() ignores attributes
a <- b <- 1:3
names(b) <- letters[1:3]
\dontrun{
expect_equal(a, b)
}
expect_equivalent(a, b)
}
\seealso{
\code{expect_setequal()} to test for set equality.

Other expectations: 
\code{\link{comparison-expectations}},
\code{\link{expect_error}()},
\code{\link{expect_length}()},
\code{\link{expect_match}()},
\code{\link{expect_message}()},
\code{\link{expect_named}()},
\code{\link{expect_null}()},
\code{\link{expect_output}()},
\code{\link{expect_reference}()},
\code{\link{expect_silent}()},
\code{\link{inheritance-expectations}},
\code{\link{logical-expectations}}
}
\concept{expectations}
