% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate-promise.R
\name{evaluate_promise}
\alias{evaluate_promise}
\title{Evaluate a promise, capturing all types of output.}
\usage{
evaluate_promise(code, print = FALSE)
}
\arguments{
\item{code}{Code to evaluate.}
}
\value{
A list containing
\item{result}{The result of the function}
\item{output}{A string containing all the output from the function}
\item{warnings}{A character vector containing the text from each warning}
\item{messages}{A character vector containing the text from each message}
}
\description{
Evaluate a promise, capturing all types of output.
}
\examples{
evaluate_promise({
  print("1")
  message("2")
  warning("3")
  4
})
}
\keyword{internal}
