% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectation.R
\name{expect}
\alias{expect}
\title{The building block of all \code{expect_} functions}
\usage{
expect(ok, failure_message, info = NULL, srcref = NULL)
}
\arguments{
\item{ok}{Was the expectation successful?}

\item{failure_message}{What message should be shown if the expectation was
not successful?}

\item{info}{Additional information. Included for backward compatibility
only and new expectations should not use it.}

\item{srcref}{Only needed in very rare circumstances where you need to
forward a srcref captured elsewhere.}
}
\description{
Use this if you are writing your own expectation. See
\code{vignette("custom-expectation")} for details
}
