% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock-object.R
\name{expect_call}
\alias{expect_call}
\title{Expectation: does the given call match the expected?}
\usage{
expect_call(mock, no, expected_call)
}
\arguments{
\item{mock}{A \code{\link{mock}} object.}

\item{no}{Number of the call made on the \code{mock} object.}

\item{expected_call}{Expected call.}
}
\description{
Expectation: does the given call match the expected?
}
\examples{
\dontrun{
expect_call(mock, 1, summary(iris))
}
}
\seealso{
Other expectations: \code{\link{comparison-expectations}},
  \code{\link{equality-expectations}},
  \code{\link{expect_equal_to_reference}},
  \code{\link{expect_length}}, \code{\link{expect_match}},
  \code{\link{expect_named}},
  \code{\link{inheritance-expectations}},
  \code{\link{logical-expectations}},
  \code{\link{output-expectations}}
}

