% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/expectations.r
\name{expect_error}
\alias{expect_error}
\title{Expectation: does expression throw an error?}
\usage{
expect_error(object, regexp = NULL, ..., info = NULL, label = NULL)
}
\arguments{
\item{object}{object to test}

\item{regexp}{optional regular expression to match. If not specified, just
asserts that expression throws some error.}

\item{...}{other arguments passed to \code{\link{matches}}}

\item{info}{extra information to be included in the message (useful when
writing tests in loops).}

\item{label}{object label. When \code{NULL}, computed from deparsed object.}
}
\description{
Expectation: does expression throw an error?
}
\examples{
f <- function() stop("My error!")
expect_error(f())
expect_error(f(), "My error!")

# You can use the arguments of grepl to control the matching
expect_error(f(), "my error!", ignore.case = TRUE)
}
\seealso{
Other expectations: \code{\link{equivalence}},
  \code{\link{expect_equal}},
  \code{\link{expect_equivalent}},
  \code{\link{expect_identical}};
  \code{\link{expect-compare}},
  \code{\link{expect_less_than}},
  \code{\link{expect_more_than}};
  \code{\link{expect_equal_to_reference}};
  \code{\link{expect_false}}, \code{\link{expect_true}};
  \code{\link{expect_is}}; \code{\link{expect_match}};
  \code{\link{expect_message}}; \code{\link{expect_named}};
  \code{\link{expect_null}}; \code{\link{expect_output}};
  \code{\link{expect_warning}};
  \code{\link{takes_less_than}}
}

