% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect-condition-forwarded.R
\name{expect_error_forwarded}
\alias{expect_error_forwarded}
\alias{expect_warning_forwarded}
\alias{expect_message_forwarded}
\title{Does code throw an error, warning, or message that is the same as another?}
\usage{
expect_error_forwarded(actual, expected, label = NULL)

expect_warning_forwarded(actual, expected, label = NULL)

expect_message_forwarded(actual, expected, label = NULL)
}
\arguments{
\item{label}{Used to customise failure messages. For expert use only.}
}
\value{
The value of the first argument
}
\description{
\code{expect_error_forwarded()}, \code{expect_warning_forwarded()}, and
\code{expect_message_forwarded()}, check that code throws an error, warning,
message, or condition with a message that matches an \code{expected} value from an
example.  These functions are useful to test that a condition matches one
from another package that you do not control.
}
\seealso{
Other expectations: 
\code{\link{comparison-expectations}},
\code{\link{equality-expectations}},
\code{\link{expect_error}()},
\code{\link{expect_length}()},
\code{\link{expect_match}()},
\code{\link{expect_named}()},
\code{\link{expect_null}()},
\code{\link{expect_output}()},
\code{\link{expect_reference}()},
\code{\link{expect_silent}()},
\code{\link{inheritance-expectations}},
\code{\link{logical-expectations}}
}
\concept{expectations}
