% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect-length.R
\name{expect_length}
\alias{expect_length}
\title{Expectation: does a vector have the specified length?}
\usage{
expect_length(object, n)
}
\arguments{
\item{object}{object to test}

\item{n}{Expected length.}
}
\description{
Expectation: does a vector have the specified length?
}
\examples{
expect_length(1, 1)
expect_length(1:10, 10)

\dontrun{
expect_length(1:10, 1)
}
}
\seealso{
Other expectations: \code{\link{comparison-expectations}},
  \code{\link{equality-expectations}},
  \code{\link{expect_equal_to_reference}},
  \code{\link{expect_match}}, \code{\link{expect_named}},
  \code{\link{inheritance-expectations}},
  \code{\link{logical-expectations}},
  \code{\link{output-expectations}}
}
