% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/expectations.r
\name{expect_match}
\alias{expect_match}
\title{Expectation: does string match regular expression?}
\usage{
expect_match(object, regexp, ..., all = TRUE, info = NULL, label = NULL)
}
\arguments{
\item{object}{object to test}

\item{regexp}{regular expression to test against}

\item{...}{For \code{matches}: other arguments passed on to
\code{\link{grepl}}. For \code{expect_match}: other arguments passed on
to \code{matches}.}

\item{all}{should all elements of actual value match \code{regexp} (TRUE),
or does only one need to match (FALSE)}

\item{info}{extra information to be included in the message (useful when
writing tests in loops).}

\item{label}{object label. When \code{NULL}, computed from deparsed object.}
}
\description{
If the object to be tested has length greater than one, all elements of
the vector must match the pattern in order to pass.
}
\examples{
expect_match("Testing is fun", "fun")
expect_match("Testing is fun", "f.n")
}
\seealso{
Other expectations: \code{\link{equivalence}},
  \code{\link{expect_equal}},
  \code{\link{expect_equivalent}},
  \code{\link{expect_identical}};
  \code{\link{expect-compare}},
  \code{\link{expect_less_than}},
  \code{\link{expect_more_than}};
  \code{\link{expect_equal_to_reference}};
  \code{\link{expect_error}}; \code{\link{expect_false}},
  \code{\link{expect_true}}; \code{\link{expect_is}};
  \code{\link{expect_message}}; \code{\link{expect_named}};
  \code{\link{expect_null}}; \code{\link{expect_output}};
  \code{\link{expect_warning}};
  \code{\link{takes_less_than}}
}

