% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/expectations.r
\name{expect_output}
\alias{expect_output}
\title{Expectation: does printed output match a regular expression?}
\usage{
expect_output(object, regexp, ..., info = NULL, label = NULL)
}
\arguments{
\item{object}{object to test}

\item{regexp}{regular expression to test against}

\item{...}{other arguments passed to \code{\link{matches}}}

\item{info}{extra information to be included in the message (useful when
writing tests in loops).}

\item{label}{object label. When \code{NULL}, computed from deparsed object.}
}
\description{
Expectation: does printed output match a regular expression?
}
\examples{
str(mtcars)
expect_output(str(mtcars), "32 obs")
expect_output(str(mtcars), "11 variables")

# You can use the arguments of grepl to control the matching
expect_output(str(mtcars), "11 VARIABLES", ignore.case = TRUE)
expect_output(str(mtcars), "$ mpg", fixed = TRUE)
}
\seealso{
Other expectations: \code{\link{equivalence}},
  \code{\link{expect_equal}},
  \code{\link{expect_equivalent}},
  \code{\link{expect_identical}};
  \code{\link{expect-compare}},
  \code{\link{expect_less_than}},
  \code{\link{expect_more_than}};
  \code{\link{expect_equal_to_reference}};
  \code{\link{expect_error}}; \code{\link{expect_false}},
  \code{\link{expect_true}}; \code{\link{expect_is}};
  \code{\link{expect_match}}; \code{\link{expect_message}};
  \code{\link{expect_named}}; \code{\link{expect_null}};
  \code{\link{expect_warning}};
  \code{\link{takes_less_than}}
}

