% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect-known.R
\name{expect_output_file}
\alias{expect_output_file}
\title{Expectations: is the output or the value equal to a known good value?}
\usage{
expect_output_file(
  object,
  file,
  update = TRUE,
  ...,
  info = NULL,
  label = NULL,
  print = FALSE,
  width = 80
)
}
\description{
\code{expect_output_file()} behaves identically to \code{\link[=expect_known_output]{expect_known_output()}}.
}
\section{3rd edition}{

\ifelse{html}{\figure{lifecycle-deprecated.svg}{options: alt='Deprecated lifecycle'}}{\strong{Deprecated}}

\code{expect_output_file()} is deprecated in the 3rd edition;
please use \code{\link[=expect_snapshot_output]{expect_snapshot_output()}} and friends instead.
}

\keyword{internal}
