% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect-length.R
\name{expect_shape}
\alias{expect_shape}
\title{Does code return an object with the specified shape?}
\description{
This is a generalization of \code{\link{expect_length}()} to test the "shape" of
  more general objects like data.frames, matrices, and arrays.
}
\usage{
expect_shape(object, ..., length, nrow, ncol, dim)
}
\arguments{
\item{object}{Object to test.

  Supports limited unquoting to make it easier to generate readable failures
  within a function or for loop. See \code{\link{quasi_label}} for more details.}

\item{...}{Ignored.}
\item{length}{Expected \code{\link{length}()} of \code{object}.}
\item{nrow}{Expected \code{\link{nrow}()} of \code{object}.}
\item{ncol}{Expected \code{\link{ncol}()} of \code{object}.}
\item{dim}{Expected \code{\link{dim}()} of \code{object}.}
}
\seealso{
\code{\link{expect_length}()} to specifically make assertions about the
  \code{\link{length}()} of a vector.

Other expectations: 
\code{\link{comparison-expectations}},
\code{\link{equality-expectations}},
\code{\link{expect_error}()},
\code{\link{expect_length}()},
\code{\link{expect_match}()},
\code{\link{expect_named}()},
\code{\link{expect_null}()},
\code{\link{expect_output}()},
\code{\link{expect_reference}()},
\code{\link{expect_silent}()},
\code{\link{inheritance-expectations}},
\code{\link{logical-expectations}}
}
\concept{expectations}
