% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect-output.R
\name{expect_silent}
\alias{expect_silent}
\title{Expectation: is the code silent?}
\usage{
expect_silent(object)
}
\arguments{
\item{object}{Object to test.

Supports limited unquoting to make it easier to generate readable failures
within a function or for loop. See \link{quasi_label} for more details.}
}
\description{
Checks that the code produces no output, messages, or warnings.
}
\examples{
expect_silent("123")

f <- function() {
  message("Hi!")
  warning("Hey!!")
  print("OY!!!")
}
\dontrun{
expect_silent(f())
}
}
