% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectations-silent.R
\name{expect_silent}
\alias{expect_silent}
\title{Expect that code has no output, messages, or warnings.}
\usage{
expect_silent(expr)
}
\arguments{
\item{expr}{Expression to evaluate}
}
\description{
Expect that code has no output, messages, or warnings.
}
\examples{
expect_silent("123")

f <- function() {
  message("Hi!")
  warning("Hey!!")
  print("OY!!!")
}
\dontrun{
expect_silent(f())
}
}
\seealso{
Other expectations: \code{\link{class-expectations}},
  \code{\link{comparison-expectations}},
  \code{\link{equality-expectations}},
  \code{\link{expect_equal_to_reference}},
  \code{\link{expect_match}}, \code{\link{expect_named}},
  \code{\link{logical-expectations}},
  \code{\link{output-expectations}},
  \code{\link{takes_less_than}}
}

