% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expectations-silent.R
\name{expect_silent}
\alias{expect_silent}
\title{Expect that code has no output, messages, or warnings.}
\usage{
expect_silent(expr)
}
\arguments{
\item{expr}{Expression to evaluate}
}
\description{
Expect that code has no output, messages, or warnings.
}
\examples{
expect_silent("123")

f <- function() {
  message("Hi!")
  warning("Hey!!")
  print("OY!!!")
}
\dontrun{
expect_silent(f())
}
}

