% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expectations-silent.R
\name{expect_silent}
\alias{expect_silent}
\title{Expect that code has no output, messages, or warnings.}
\usage{
expect_silent(expr)
}
\arguments{
\item{expr}{Expression to evaluate}
}
\description{
Expect that code has no output, messages, or warnings.
}
\examples{
expect_silent("123")

f <- function() {
  message("Hi!")
  warning("Hey!!")
  print("OY!!!")
}
\dontrun{
expect_silent(f())
}
}
\seealso{
Other expectations: \code{\link{equivalence}},
  \code{\link{expect_equal}},
  \code{\link{expect_equivalent}},
  \code{\link{expect_identical}};
  \code{\link{expect-compare}}, \code{\link{expect_gt}},
  \code{\link{expect_gte}}, \code{\link{expect_less_than}},
  \code{\link{expect_lt}}, \code{\link{expect_lte}},
  \code{\link{expect_more_than}};
  \code{\link{expect_equal_to_reference}};
  \code{\link{expect_error}}, \code{\link{expect_match}},
  \code{\link{expect_message}},
  \code{\link{expect_output}},
  \code{\link{expect_warning}},
  \code{\link{matching-expectations}};
  \code{\link{expect_false}}, \code{\link{expect_true}};
  \code{\link{expect_is}}; \code{\link{expect_named}};
  \code{\link{expect_null}}; \code{\link{takes_less_than}}
}

