% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect-self-test.R
\name{expect_success}
\alias{expect_success}
\alias{expect_failure}
\title{Tools for testing expectations}
\usage{
expect_success(expr)

expect_failure(expr, message = NULL, ...)
}
\arguments{
\item{expr}{Expression that evaluates a single expectation.}

\item{message}{Check that the failure message matches this regexp.}

\item{...}{Other arguments passed on to \code{\link[=expect_match]{expect_match()}}.}
}
\description{
Use these expectations to test other expectations.
}
\keyword{internal}
