% Generated by roxygen2 (4.0.0): do not edit by hand
\name{expectation}
\alias{expectation}
\alias{format.expectation}
\alias{is.expectation}
\alias{print.expectation}
\title{Expectation class.}
\usage{
expectation(passed, failure_msg, success_msg = "unknown", pending = FALSE,
  pending_msg = "unknown")

is.expectation(x)
}
\arguments{
  \item{passed}{a single logical value indicating whether
  the test passed (\code{TRUE}), failed (\code{FALSE}), or
  threw an error (\code{NA})}

  \item{failure_msg}{A text description of failure}

  \item{success_msg}{A text description of success}

  \item{pending}{single logical value: TRUE iff test is
  pending}

  \item{pending_msg}{A text description of pending}

  \item{x}{object to test for class membership}
}
\description{
Any expectation should return objects of this class - see the built in
expectations for details.
}
\keyword{internal}

