% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectation.R
\name{expectation}
\alias{expectation}
\alias{is.expectation}
\title{Construct an expectation object}
\usage{
expectation(type, message, srcref = NULL, trace = NULL, env = NULL)

is.expectation(x)
}
\arguments{
\item{type}{Expectation type. Must be one of "success", "failure", "error",
"skip", "warning".}

\item{message}{Message describing test failure}

\item{srcref}{Optional \code{srcref} giving location of test.}

\item{trace}{An optional backtrace created by \code{\link[rlang:trace_back]{rlang::trace_back()}}.
When supplied, the expectation is displayed with the backtrace.}

\item{env}{The environment of the \code{expect_} function. You should
only pass this when you're calling \code{expect()} from an internal
helper. This environment is passed to \code{\link[rlang:trace_back]{rlang::trace_back()}} as
\code{bottom} argument, in order to remove the uninformative testthat
context from backtraces when an unexpected error occurs during
evaluation of \code{ok}. Has no effect if \code{trace} is supplied.}

\item{x}{object to test for class membership}
}
\description{
For advanced use only. If you are creating your own expectation, you should
call \code{\link[=expect]{expect()}} instead. See \code{vignette("custom-expectation")} for more
details.
}
\keyword{internal}
