\name{expectation}
\title{Expectation class.}
\usage{expectation(passed, message)}

\description{
  Expectation class.
}

\details{
  Any expectation should return objects of this class - see
  the built in expectations for details.
}
\alias{expectation}
\alias{print.expectation}
\alias{format.expectation}
\keyword{internal}
