% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectation.R
\name{expectation}
\alias{expectation}
\alias{is.expectation}
\title{Construct an expectation object}
\usage{
expectation(type, message, srcref = NULL)

is.expectation(x)
}
\arguments{
\item{x}{object to test for class membership}
}
\description{
For advanced use only.
}
\keyword{internal}
