% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fail}
\alias{fail}
\title{A default expectation that always fails.}
\usage{
fail(message = "Failure has been forced.")
}
\arguments{
\item{message}{a string to display.}
}
\description{
The fail function forces a test to fail.  This is useful if you want to
test a pre-condition '
}
\examples{
\dontrun{
test_that("this test fails", fail())
}
}

