% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect-that.R
\name{fail}
\alias{fail}
\alias{succeed}
\title{Default expectations that always succeed or fail.}
\usage{
fail(message = "Failure has been forced", info = NULL)

succeed(message = "Success has been forced", info = NULL)
}
\arguments{
\item{message}{a string to display.}

\item{info}{Character vector continuing additional information. Included
for backward compatibility only and new expectations should not use it.}
}
\description{
These allow you to manually trigger success or failure. Failure is
particularly useful to a pre-condition or mark a test as not yet
implemented.
}
\examples{
\dontrun{
test_that("this test fails", fail())
test_that("this test succeeds", succeed())
}
}
