% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-fail.R
\docType{data}
\name{FailReporter}
\alias{FailReporter}
\title{Test reporter: fail at end.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FailReporter
}
\description{
This reporter will simply throw an error if any of the tests failed. It is
best combined with another reporter, such as the
\code{\link{SummaryReporter}}.
}
\keyword{datasets}

