% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-fail.R
\name{FailReporter}
\alias{FailReporter}
\title{Test reporter: fail at end.}
\description{
This reporter will simply throw an error if any of the tests failed. It is
best combined with another reporter, such as the
\link{SummaryReporter}.
}
\seealso{
Other reporters: 
\code{\link{CheckReporter}},
\code{\link{DebugReporter}},
\code{\link{ListReporter}},
\code{\link{LocationReporter}},
\code{\link{MinimalReporter}},
\code{\link{MultiReporter}},
\code{\link{ProgressReporter}},
\code{\link{Reporter}},
\code{\link{RstudioReporter}},
\code{\link{SilentReporter}},
\code{\link{StopReporter}},
\code{\link{SummaryReporter}},
\code{\link{TapReporter}},
\code{\link{TeamcityReporter}}
}
\concept{reporters}
