% Generated by roxygen2 (4.0.2): do not edit by hand
\name{find_reporter}
\alias{find_reporter}
\title{Find reporter object given name}
\usage{
find_reporter(reporter)
}
\arguments{
\item{reporter}{name of reporter}
}
\description{
If not found, will return informative error message
}
\keyword{internal}

