% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reporter-zzz.r
\name{find_reporter}
\alias{find_reporter}
\title{Find reporter object given name or object.}
\usage{
find_reporter(reporter)
}
\arguments{
\item{reporter}{name of reporter}
}
\description{
If not found, will return informative error message.
Will return null if given NULL.
}
\examples{
testthat:::find_reporter("minimal")
testthat:::find_reporter("summary+fail")
}
\keyword{internal}

