\name{find_reporter}
\alias{find_reporter}
\title{Find reporter object given name...}
\usage{find_reporter(reporter)}

\description{
  Find reporter object given name
}

\details{
  If not found, will return informative error message
}
\keyword{internal}
\arguments{
  \item{reporter}{name of reporter}
}
