% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-zzz.R
\name{find_reporter}
\alias{find_reporter}
\title{Find reporter object given name or object.}
\usage{
find_reporter(reporter)
}
\arguments{
\item{reporter}{name of reporter(s), or reporter object(s)}
}
\description{
If not found, will return informative error message.
Pass a character vector to create a \link{MultiReporter} composed
of individual reporters.
Will return null if given NULL.
}
\keyword{internal}
