% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-files.R
\name{find_test_scripts}
\alias{find_test_scripts}
\title{Find the test files.}
\usage{
find_test_scripts(path, filter = NULL, invert = FALSE, ...)
}
\arguments{
\item{path}{path to tests}

\item{filter}{cf \code{\link[=test_dir]{test_dir()}}}

\item{invert}{If \sQuote{TRUE} return files which do \emph{not} match.}

\item{...}{Additional arguments passed to \code{\link[=grepl]{grepl()}} to control filtering.}
}
\value{
the test file paths
}
\description{
Find the test files.
}
\keyword{internal}
