% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-files.R
\name{find_test_scripts}
\alias{find_test_scripts}
\title{Find test files}
\usage{
find_test_scripts(path, filter = NULL, invert = FALSE, ..., full.names = TRUE)
}
\arguments{
\item{path}{path to tests}

\item{filter}{If not \code{NULL}, only tests with file names matching this
regular expression will be executed. Matching is performed on the file
name after it's stripped of \code{"test-"} and \code{".R"}.}

\item{invert}{If \code{TRUE} return files which \strong{don't} match.}

\item{...}{Additional arguments passed to \code{\link[=grepl]{grepl()}} to control filtering.}
}
\value{
A character vector of paths
}
\description{
Find test files
}
\keyword{internal}
