% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test-files.r
\name{find_test_scripts}
\alias{find_test_scripts}
\title{Find the test files.}
\usage{
find_test_scripts(path, filter = NULL, ...)
}
\arguments{
\item{path}{path to tests}

\item{filter}{cf \code{\link{test_dir}}}

\item{...}{Additional arguments passed to \code{grepl} to control filtering.}
}
\value{
the test file paths
}
\description{
Find the test files.
}
\keyword{internal}

