% Generated by roxygen2 (4.0.2): do not edit by hand
\name{is_equivalent_to}
\alias{expect_equivalent}
\alias{is_equivalent_to}
\title{Expectation: is the object equivalent to a value?
This expectation tests for equivalency: are two objects equal once their
attributes have been removed.}
\usage{
is_equivalent_to(expected, label = NULL)

expect_equivalent(object, expected, info = NULL, label = NULL,
  expected.label = NULL)
}
\arguments{
\item{expected}{Expected value}

\item{label}{For full form, label of expected object used in error
messages. Useful to override default (deparsed expected expression) when
doing tests in a loop.  For short cut form, object label. When
\code{NULL}, computed from deparsed object.}

\item{object}{object to test}

\item{info}{extra information to be included in the message (useful when
writing tests in loops).}

\item{expected.label}{Equivalent of \code{label} for shortcut form.}
}
\description{
Expectation: is the object equivalent to a value?
This expectation tests for equivalency: are two objects equal once their
attributes have been removed.
}
\examples{
a <- b <- 1:3
names(b) <- letters[1:3]
expect_equivalent(a, b)
}
\seealso{
Other expectations: \code{\link{equals_reference}},
  \code{\link{expect_equal_to_reference}};
  \code{\link{equals}}, \code{\link{expect_equal}};
  \code{\link{expect-compare}},
  \code{\link{expect_less_than}},
  \code{\link{expect_more_than}},
  \code{\link{is_less_than}}, \code{\link{is_more_than}};
  \code{\link{expect_error}}, \code{\link{throws_error}};
  \code{\link{expect_false}}, \code{\link{expect_true}},
  \code{\link{is_false}}, \code{\link{is_true}};
  \code{\link{expect_identical}},
  \code{\link{is_identical_to}}; \code{\link{expect_is}},
  \code{\link{is_a}}; \code{\link{expect_match}},
  \code{\link{matches}}; \code{\link{expect_message}},
  \code{\link{shows_message}}; \code{\link{expect_named}},
  \code{\link{has_names}}; \code{\link{expect_null}},
  \code{\link{is_null}}; \code{\link{expect_output}},
  \code{\link{prints_text}}; \code{\link{expect_warning}},
  \code{\link{gives_warning}};
  \code{\link{takes_less_than}}
}

