\name{is_equivalent_to}
\alias{is_equivalent_to}
\title{Expectation: is the object equivalent to a value?}
\usage{is_equivalent_to(expected)}

\description{
  Expectation: is the object equivalent to a value? This expectation
  tests for equivalency: are two objects equal once their attributes
  have been removed.
}
\arguments{
  \item{expected}{expected value}
}
