% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-directory.R
\name{is_testing}
\alias{is_testing}
\alias{testing_package}
\title{Determine testing status}
\usage{
is_testing()

testing_package()
}
\description{
Use \code{is_testing()} to determine if code is being run as part of a test and
\code{testing_package()} to retrieve the name of the package being tested. You
can also check the underlying env var directly
\code{identical(Sys.getenv("TESTTHAT"), "true")} to avoid creating a run-time
dependency on testthat.
}
