% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-env.R
\name{is_testing}
\alias{is_testing}
\alias{is_parallel}
\alias{testing_package}
\title{Determine testing status}
\usage{
is_testing()

is_parallel()

testing_package()
}
\description{
\itemize{
\item \code{is_testing()} determine if code is being run as part of a test
\item \code{is_parallel()} if the test is being run in parallel.
\item \code{testing_package()} gives name of the package being tested.
}

These are thin wrappers that retrieve the values of environment variables.
You will usually want to avoid a runtime dependency on testthat, i.e. you
won't want to use testthat in the code below \verb{R/}. In that case, you should
inline the source of these functions into your package, in whatever location
you are using for any other test helper functions.
}
