% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-junit.R
\name{JunitReporter}
\alias{JunitReporter}
\title{Test reporter: summary of errors in jUnit XML format.}
\description{
This reporter includes detailed results about each test and summaries,
written to a file (or stdout) in jUnit XML format. This can be read by
the Jenkins Continuous Integration System to report on a dashboard etc.
Requires the \emph{xml2} package.
}
\details{
To fit into the jUnit structure, context() becomes the \verb{<testsuite>}
name as well as the base of the \verb{<testcase> classname}. The
test_that() name becomes the rest of the \verb{<testcase> classname}.
The deparsed expect_that() call becomes the \verb{<testcase>} name.
On failure, the message goes into the \verb{<failure>} node message
argument (first line only) and into its text content (full message).

Execution time and some other details are also recorded.

References for the jUnit XML format:
\url{http://llg.cubic.org/docs/junit/}
}
\section{Super class}{
\code{\link[testthat:Reporter]{testthat::Reporter}} -> \code{JunitReporter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-elapsed_time}{\code{JunitReporter$elapsed_time()}}
\item \href{#method-reset_suite}{\code{JunitReporter$reset_suite()}}
\item \href{#method-start_reporter}{\code{JunitReporter$start_reporter()}}
\item \href{#method-start_file}{\code{JunitReporter$start_file()}}
\item \href{#method-start_test}{\code{JunitReporter$start_test()}}
\item \href{#method-start_context}{\code{JunitReporter$start_context()}}
\item \href{#method-end_context}{\code{JunitReporter$end_context()}}
\item \href{#method-add_result}{\code{JunitReporter$add_result()}}
\item \href{#method-end_reporter}{\code{JunitReporter$end_reporter()}}
\item \href{#method-clone}{\code{JunitReporter$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id=".end_context">}\href{../../testthat/html/Reporter.html#method-.end_context}{\code{testthat::Reporter$.end_context()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id=".start_context">}\href{../../testthat/html/Reporter.html#method-.start_context}{\code{testthat::Reporter$.start_context()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="cat_line">}\href{../../testthat/html/Reporter.html#method-cat_line}{\code{testthat::Reporter$cat_line()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="cat_tight">}\href{../../testthat/html/Reporter.html#method-cat_tight}{\code{testthat::Reporter$cat_tight()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="end_file">}\href{../../testthat/html/Reporter.html#method-end_file}{\code{testthat::Reporter$end_file()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="end_test">}\href{../../testthat/html/Reporter.html#method-end_test}{\code{testthat::Reporter$end_test()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="initialize">}\href{../../testthat/html/Reporter.html#method-initialize}{\code{testthat::Reporter$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="is_full">}\href{../../testthat/html/Reporter.html#method-is_full}{\code{testthat::Reporter$is_full()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="rule">}\href{../../testthat/html/Reporter.html#method-rule}{\code{testthat::Reporter$rule()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-elapsed_time"></a>}}
\subsection{Method \code{elapsed_time()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JunitReporter$elapsed_time()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset_suite"></a>}}
\subsection{Method \code{reset_suite()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JunitReporter$reset_suite()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-start_reporter"></a>}}
\subsection{Method \code{start_reporter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JunitReporter$start_reporter()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-start_file"></a>}}
\subsection{Method \code{start_file()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JunitReporter$start_file(file)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-start_test"></a>}}
\subsection{Method \code{start_test()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JunitReporter$start_test(context, test)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-start_context"></a>}}
\subsection{Method \code{start_context()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JunitReporter$start_context(context)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-end_context"></a>}}
\subsection{Method \code{end_context()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JunitReporter$end_context(context)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_result"></a>}}
\subsection{Method \code{add_result()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JunitReporter$add_result(context, test, result)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-end_reporter"></a>}}
\subsection{Method \code{end_reporter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JunitReporter$end_reporter()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JunitReporter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
