% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{ListReporter-class}
\alias{ListReporter}
\alias{ListReporter-class}
\title{List reporter: gather all test results along with elapsed time and
file information.}
\arguments{
\item{...}{Arguments used to initialise class}
}
\description{
This reporter gathers all results, adding additional information such as
test elapsed time, and test filename if available. Very useful for reporting.
}
\keyword{debugging}

