\name{ListReporter}
\alias{ListReporter}
\title{List reporter: gather all test results along with elapsed time and
file information.}
\usage{
  ListReporter(...)
}
\arguments{
  \item{...}{Arguments used to initialise class}
}
\description{
  This reporter gathers all results, adding additional
  information such as test elapsed time, and test filename
  if available. Very useful for reporting.
}
\keyword{debugging}

