% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-list.R
\name{ListReporter}
\alias{ListReporter}
\title{List reporter: gather all test results along with elapsed time and
file information.}
\description{
This reporter gathers all results, adding additional information such as
test elapsed time, and test filename if available. Very useful for reporting.
}
\seealso{
Other reporters: 
\code{\link{CheckReporter}},
\code{\link{DebugReporter}},
\code{\link{FailReporter}},
\code{\link{JunitReporter}},
\code{\link{LocationReporter}},
\code{\link{MinimalReporter}},
\code{\link{MultiReporter}},
\code{\link{ProgressReporter}},
\code{\link{RStudioReporter}},
\code{\link{Reporter}},
\code{\link{SilentReporter}},
\code{\link{StopReporter}},
\code{\link{SummaryReporter}},
\code{\link{TapReporter}},
\code{\link{TeamcityReporter}}
}
\concept{reporters}
