% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-llm.R
\name{LlmReporter}
\alias{LlmReporter}
\title{Report test progress for LLMs}
\description{
\code{LlmReporter} is designed for use with Large Language Models (LLMs).
It only reports problems (warnings, skips, errors, and failures) as they
occur, only reporting passing tests at the end.
}
\seealso{
Other reporters: 
\code{\link{CheckReporter}},
\code{\link{DebugReporter}},
\code{\link{FailReporter}},
\code{\link{JunitReporter}},
\code{\link{ListReporter}},
\code{\link{LocationReporter}},
\code{\link{MinimalReporter}},
\code{\link{MultiReporter}},
\code{\link{ProgressReporter}},
\code{\link{RStudioReporter}},
\code{\link{Reporter}},
\code{\link{SilentReporter}},
\code{\link{SlowReporter}},
\code{\link{StopReporter}},
\code{\link{SummaryReporter}},
\code{\link{TapReporter}},
\code{\link{TeamcityReporter}}
}
\concept{reporters}
