% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot-reporter.R
\name{local_snapshotter}
\alias{local_snapshotter}
\title{Instantiate local snapshotting context}
\usage{
local_snapshotter(snap_dir = "_snaps", cleanup = FALSE, .env = parent.frame())
}
\description{
Needed if you want to run snapshot tests outside of the usual testthat
framework For expert use only.
}
\keyword{internal}
