% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-location.R
\name{LocationReporter}
\alias{LocationReporter}
\title{Test reporter: location}
\description{
This reporter simply prints the location of every expectation and error.
This is useful if you're trying to figure out the source of a segfault,
or you want to figure out which code triggers a C/C++ breakpoint
}
\seealso{
Other reporters: 
\code{\link{CheckReporter}},
\code{\link{DebugReporter}},
\code{\link{FailReporter}},
\code{\link{ListReporter}},
\code{\link{MinimalReporter}},
\code{\link{MultiReporter}},
\code{\link{ProgressReporter}},
\code{\link{Reporter}},
\code{\link{RstudioReporter}},
\code{\link{SilentReporter}},
\code{\link{StopReporter}},
\code{\link{SummaryReporter}},
\code{\link{TapReporter}},
\code{\link{TeamcityReporter}}
}
\concept{reporters}
\section{Super class}{
\code{\link[testthat:Reporter]{testthat::Reporter}} -> \code{LocationReporter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-start_test}{\code{LocationReporter$start_test()}}
\item \href{#method-add_result}{\code{LocationReporter$add_result()}}
\item \href{#method-end_test}{\code{LocationReporter$end_test()}}
\item \href{#method-clone}{\code{LocationReporter$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id=".end_context">}\href{../../testthat/html/Reporter.html#method-.end_context}{\code{testthat::Reporter$.end_context()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id=".start_context">}\href{../../testthat/html/Reporter.html#method-.start_context}{\code{testthat::Reporter$.start_context()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="cat_line">}\href{../../testthat/html/Reporter.html#method-cat_line}{\code{testthat::Reporter$cat_line()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="cat_tight">}\href{../../testthat/html/Reporter.html#method-cat_tight}{\code{testthat::Reporter$cat_tight()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="end_context">}\href{../../testthat/html/Reporter.html#method-end_context}{\code{testthat::Reporter$end_context()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="end_file">}\href{../../testthat/html/Reporter.html#method-end_file}{\code{testthat::Reporter$end_file()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="end_reporter">}\href{../../testthat/html/Reporter.html#method-end_reporter}{\code{testthat::Reporter$end_reporter()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="initialize">}\href{../../testthat/html/Reporter.html#method-initialize}{\code{testthat::Reporter$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="is_full">}\href{../../testthat/html/Reporter.html#method-is_full}{\code{testthat::Reporter$is_full()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="rule">}\href{../../testthat/html/Reporter.html#method-rule}{\code{testthat::Reporter$rule()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="start_context">}\href{../../testthat/html/Reporter.html#method-start_context}{\code{testthat::Reporter$start_context()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="start_file">}\href{../../testthat/html/Reporter.html#method-start_file}{\code{testthat::Reporter$start_file()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="start_reporter">}\href{../../testthat/html/Reporter.html#method-start_reporter}{\code{testthat::Reporter$start_reporter()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-start_test"></a>}}
\subsection{Method \code{start_test()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocationReporter$start_test(context, test)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_result"></a>}}
\subsection{Method \code{add_result()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocationReporter$add_result(context, test, result)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-end_test"></a>}}
\subsection{Method \code{end_test()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocationReporter$end_test(context, test)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocationReporter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
