% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-location.R
\name{LocationReporter}
\alias{LocationReporter}
\title{Test reporter: location}
\description{
This reporter simply prints the location of every expectation and error.
This is useful if you're trying to figure out the source of a segfault,
or you want to figure out which code triggers a C/C++ breakpoint
}
\seealso{
Other reporters: 
\code{\link{CheckReporter}},
\code{\link{DebugReporter}},
\code{\link{FailReporter}},
\code{\link{JunitReporter}},
\code{\link{ListReporter}},
\code{\link{MinimalReporter}},
\code{\link{MultiReporter}},
\code{\link{ProgressReporter}},
\code{\link{RStudioReporter}},
\code{\link{Reporter}},
\code{\link{SilentReporter}},
\code{\link{StopReporter}},
\code{\link{SummaryReporter}},
\code{\link{TapReporter}},
\code{\link{TeamcityReporter}}
}
\concept{reporters}
