\docType{data}
\name{MinimalReporter}
\alias{MinimalReporter}
\title{Test reporter: minimal.}
\format{Reference class 'refObjectGenerator' [package "methods"] with 2 fields
 and 17 methods, of which 6 are possibly relevant:
   accessors, fields, help, lock, methods, new}
\usage{
  MinimalReporter
}
\description{
  The minimal test reporter provides the absolutely minimum
  amount of information: whether each expectation has
  succeeded, failed or experienced an error.  If you want
  to find out what the failures and errors actually were,
  you'll need to run a more informative test reporter.
}
\keyword{dataset}
\keyword{debugging}

