\name{MinimalReporter}
\alias{MinimalReporter}
\title{Test reporter: minimal.}
\description{
  The minimal test reporter provides the absolutely minimum
  amount of information: whether each expectation has
  succeeded, failed or experienced an error.  If you want
  to find out what the failures and errors actually were,
  you'll need to run a more informative test reporter.
}
\keyword{debugging}

