% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-minimal.R
\docType{data}
\name{MinimalReporter}
\alias{MinimalReporter}
\title{Test reporter: minimal.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MinimalReporter
}
\description{
The minimal test reporter provides the absolutely minimum amount of
information: whether each expectation has succeeded, failed or experienced
an error.  If you want to find out what the failures and errors actually
were, you'll need to run a more informative test reporter.
}
\keyword{datasets}

