% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock-object.R
\name{mock}
\alias{mock}
\title{Create a mock object.}
\usage{
mock(returns = list())
}
\arguments{
\item{returns}{Values returned upon subsequent calls.}
}
\description{
Create a mock object.
}
\examples{
\dontrun{
  m <- mock(returns = list(1))
  with_mock(summary = m, {
    expect_equal(summary(iris), 1)
    expect_call(m, 1, summary(iris))
  })
}

}

