% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reporter-multi.r
\docType{class}
\name{MultiReporter-class}
\alias{MultiReporter}
\alias{MultiReporter-class}
\title{Multi reporter: combine several reporters in one.}
\arguments{
\item{...}{Arguments used to initialise class}
}
\description{
This reporter is useful to use several reporters at the same time, e.g.
adding a custom reporter without removing the current one.
}
\keyword{debugging}

