% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-multi.R
\name{MultiReporter}
\alias{MultiReporter}
\title{Multi reporter: combine several reporters in one.}
\description{
This reporter is useful to use several reporters at the same time, e.g.
adding a custom reporter without removing the current one.
}
\seealso{
Other reporters: 
\code{\link{CheckReporter}},
\code{\link{DebugReporter}},
\code{\link{FailReporter}},
\code{\link{ListReporter}},
\code{\link{LocationReporter}},
\code{\link{MinimalReporter}},
\code{\link{ProgressReporter}},
\code{\link{Reporter}},
\code{\link{RstudioReporter}},
\code{\link{SilentReporter}},
\code{\link{StopReporter}},
\code{\link{SummaryReporter}},
\code{\link{TapReporter}},
\code{\link{TeamcityReporter}}
}
\concept{reporters}
\section{Super class}{
\code{\link[testthat:Reporter]{testthat::Reporter}} -> \code{MultiReporter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MultiReporter$new()}}
\item \href{#method-start_reporter}{\code{MultiReporter$start_reporter()}}
\item \href{#method-start_file}{\code{MultiReporter$start_file()}}
\item \href{#method-start_context}{\code{MultiReporter$start_context()}}
\item \href{#method-start_test}{\code{MultiReporter$start_test()}}
\item \href{#method-add_result}{\code{MultiReporter$add_result()}}
\item \href{#method-end_test}{\code{MultiReporter$end_test()}}
\item \href{#method-end_context}{\code{MultiReporter$end_context()}}
\item \href{#method-end_reporter}{\code{MultiReporter$end_reporter()}}
\item \href{#method-end_file}{\code{MultiReporter$end_file()}}
\item \href{#method-clone}{\code{MultiReporter$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id=".end_context">}\href{../../testthat/html/Reporter.html#method-.end_context}{\code{testthat::Reporter$.end_context()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id=".start_context">}\href{../../testthat/html/Reporter.html#method-.start_context}{\code{testthat::Reporter$.start_context()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="cat_line">}\href{../../testthat/html/Reporter.html#method-cat_line}{\code{testthat::Reporter$cat_line()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="cat_tight">}\href{../../testthat/html/Reporter.html#method-cat_tight}{\code{testthat::Reporter$cat_tight()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="is_full">}\href{../../testthat/html/Reporter.html#method-is_full}{\code{testthat::Reporter$is_full()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="rule">}\href{../../testthat/html/Reporter.html#method-rule}{\code{testthat::Reporter$rule()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiReporter$new(reporters = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-start_reporter"></a>}}
\subsection{Method \code{start_reporter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiReporter$start_reporter()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-start_file"></a>}}
\subsection{Method \code{start_file()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiReporter$start_file(filename)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-start_context"></a>}}
\subsection{Method \code{start_context()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiReporter$start_context(context)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-start_test"></a>}}
\subsection{Method \code{start_test()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiReporter$start_test(context, test)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_result"></a>}}
\subsection{Method \code{add_result()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiReporter$add_result(context, test, result)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-end_test"></a>}}
\subsection{Method \code{end_test()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiReporter$end_test(context, test)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-end_context"></a>}}
\subsection{Method \code{end_context()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiReporter$end_context(context)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-end_reporter"></a>}}
\subsection{Method \code{end_reporter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiReporter$end_reporter()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-end_file"></a>}}
\subsection{Method \code{end_file()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiReporter$end_file()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiReporter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
