% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-multi.R
\docType{data}
\name{MultiReporter}
\alias{MultiReporter}
\title{Multi reporter: combine several reporters in one.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MultiReporter
}
\description{
This reporter is useful to use several reporters at the same time, e.g.
adding a custom reporter without removing the current one.
}
\keyword{datasets}

