% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect-that.R
\name{not}
\alias{not}
\title{Negate an expectation}
\usage{
not(f)
}
\arguments{
\item{f}{an existing expectation function}
}
\description{
This negates an expectation, making it possible to express that you
want the opposite of a standard expectation. This function is deprecated
and will be removed in a future version.
}
\keyword{internal}
