% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expectations-equality.R, R/expectations-matches.R, R/expectations-old.R, R/expectations.r
\name{equals}
\alias{equals}
\alias{equals_reference}
\alias{gives_warning}
\alias{has_names}
\alias{is_a}
\alias{is_equivalent_to}
\alias{is_false}
\alias{is_identical_to}
\alias{is_less_than}
\alias{is_more_than}
\alias{is_null}
\alias{is_true}
\alias{matches}
\alias{oldskool}
\alias{prints_text}
\alias{shows_message}
\alias{throws_error}
\title{Old-style expectations.}
\usage{
equals(expected, label = NULL, ...)

is_equivalent_to(expected, label = NULL)

is_identical_to(expected, label = NULL)

equals_reference(file, label = NULL, ...)

matches(regexp, all = TRUE, ...)

prints_text(regexp, ...)

throws_error(regexp = NULL, ...)

gives_warning(regexp = NULL, all = FALSE, ...)

shows_message(regexp = NULL, all = FALSE, ...)

is_a(class)

is_true()

is_false()

is_null()

has_names(expected, ignore.order = FALSE, ignore.case = FALSE)

is_less_than(expected, label = NULL, ...)

is_more_than(expected, label = NULL, ...)
}
\description{
Initial testthat used a style of testing that looked like
\code{expect_that(a, equals(b)))} this allowed expectations to read like
English sentences, but was verbose and a bit too cutesy. This style
will continue to work but has been soft-deprecated - it is no longer
documented, and new expectations will only use the new style
\code{expect_equal(a, b)}.
}
\keyword{internal}

