% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-progress.R
\name{ProgressReporter}
\alias{ProgressReporter}
\title{Test reporter: interactive progress bar of errors.}
\description{
This reporter is a reimagining of \link{SummaryReporter} designed to make the
most information available up front, while taking up less space overall. It
is the default reporting reporter used by \code{\link[=test_dir]{test_dir()}} and \code{\link[=test_file]{test_file()}}.
}
\details{
As an additional benefit, this reporter will praise you from time-to-time
if all your tests pass.
}
\seealso{
Other reporters: 
\code{\link{CheckReporter}},
\code{\link{DebugReporter}},
\code{\link{FailReporter}},
\code{\link{ListReporter}},
\code{\link{LocationReporter}},
\code{\link{MinimalReporter}},
\code{\link{MultiReporter}},
\code{\link{Reporter}},
\code{\link{RstudioReporter}},
\code{\link{SilentReporter}},
\code{\link{StopReporter}},
\code{\link{SummaryReporter}},
\code{\link{TapReporter}},
\code{\link{TeamcityReporter}}
}
\concept{reporters}
\section{Super class}{
\code{\link[testthat:Reporter]{testthat::Reporter}} -> \code{ProgressReporter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ProgressReporter$new()}}
\item \href{#method-is_full}{\code{ProgressReporter$is_full()}}
\item \href{#method-start_reporter}{\code{ProgressReporter$start_reporter()}}
\item \href{#method-start_file}{\code{ProgressReporter$start_file()}}
\item \href{#method-start_test}{\code{ProgressReporter$start_test()}}
\item \href{#method-start_context}{\code{ProgressReporter$start_context()}}
\item \href{#method-show_header}{\code{ProgressReporter$show_header()}}
\item \href{#method-show_status}{\code{ProgressReporter$show_status()}}
\item \href{#method-end_context}{\code{ProgressReporter$end_context()}}
\item \href{#method-add_result}{\code{ProgressReporter$add_result()}}
\item \href{#method-end_reporter}{\code{ProgressReporter$end_reporter()}}
\item \href{#method-should_update}{\code{ProgressReporter$should_update()}}
\item \href{#method-clone}{\code{ProgressReporter$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id=".end_context">}\href{../../testthat/html/Reporter.html#method-.end_context}{\code{testthat::Reporter$.end_context()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id=".start_context">}\href{../../testthat/html/Reporter.html#method-.start_context}{\code{testthat::Reporter$.start_context()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="cat_line">}\href{../../testthat/html/Reporter.html#method-cat_line}{\code{testthat::Reporter$cat_line()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="cat_tight">}\href{../../testthat/html/Reporter.html#method-cat_tight}{\code{testthat::Reporter$cat_tight()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="end_file">}\href{../../testthat/html/Reporter.html#method-end_file}{\code{testthat::Reporter$end_file()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="end_test">}\href{../../testthat/html/Reporter.html#method-end_test}{\code{testthat::Reporter$end_test()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="rule">}\href{../../testthat/html/Reporter.html#method-rule}{\code{testthat::Reporter$rule()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProgressReporter$new(
  show_praise = TRUE,
  max_failures = getOption("testthat.progress.max_fails", 10L),
  min_time = 0.1,
  update_interval = 0.1,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_full"></a>}}
\subsection{Method \code{is_full()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProgressReporter$is_full()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-start_reporter"></a>}}
\subsection{Method \code{start_reporter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProgressReporter$start_reporter(context)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-start_file"></a>}}
\subsection{Method \code{start_file()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProgressReporter$start_file(file)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-start_test"></a>}}
\subsection{Method \code{start_test()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProgressReporter$start_test(context, test)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-start_context"></a>}}
\subsection{Method \code{start_context()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProgressReporter$start_context(context)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-show_header"></a>}}
\subsection{Method \code{show_header()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProgressReporter$show_header()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-show_status"></a>}}
\subsection{Method \code{show_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProgressReporter$show_status(complete = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-end_context"></a>}}
\subsection{Method \code{end_context()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProgressReporter$end_context(context)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_result"></a>}}
\subsection{Method \code{add_result()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProgressReporter$add_result(context, test, result)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-end_reporter"></a>}}
\subsection{Method \code{end_reporter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProgressReporter$end_reporter()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-should_update"></a>}}
\subsection{Method \code{should_update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProgressReporter$should_update()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProgressReporter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
