% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-progress.R
\name{ProgressReporter}
\alias{ProgressReporter}
\title{Test reporter: interactive progress bar of errors.}
\description{
This reporter is a reimagining of \link{SummaryReporter} designed to make the
most information available up front, while taking up less space overall. It
is the default reporting reporter used by \code{\link[=test_dir]{test_dir()}} and \code{\link[=test_file]{test_file()}}.
}
\details{
As an additional benefit, this reporter will praise you from time-to-time
if all your tests pass.
}
\seealso{
Other reporters: 
\code{\link{CheckReporter}},
\code{\link{DebugReporter}},
\code{\link{FailReporter}},
\code{\link{ListReporter}},
\code{\link{LocationReporter}},
\code{\link{MinimalReporter}},
\code{\link{MultiReporter}},
\code{\link{Reporter}},
\code{\link{RstudioReporter}},
\code{\link{SilentReporter}},
\code{\link{StopReporter}},
\code{\link{SummaryReporter}},
\code{\link{TapReporter}},
\code{\link{TeamcityReporter}}
}
\concept{reporters}
