\docType{data}
\name{Reporter}
\alias{Reporter}
\title{Stub object for managing a reporter of tests.}
\format{Reference class 'refObjectGenerator' [package "methods"] with 2 fields
 and 17 methods, of which 6 are possibly relevant:
   accessors, fields, help, lock, methods, new}
\usage{
  Reporter
}
\description{
  Do not clone directly from this object - children should
  implement all methods.
}
\keyword{dataset}
\keyword{internal.}

