\name{Reporter}
\alias{Reporter}
\alias{Reporter-class}
\title{Stub object for managing a reporter of tests.}
\usage{
  Reporter(...)
}
\arguments{
  \item{...}{Arguments used to initialise class}
}
\description{
  Do not clone directly from this object - children should
  implement all methods.
}
\keyword{internal.}

