% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter.R
\name{Reporter}
\alias{Reporter}
\title{Manage test reporting}
\description{
The job of a reporter is to aggregate the results from files, tests, and
expectations and display them in an informative way. Every testtthat function
that runs multiple tests provides a \code{reporter} argument which you can
use to override the default (which is selected by \code{\link[=default_reporter]{default_reporter()}}).
}
\details{
You only need to use this \code{Reporter} object directly if you are creating
a new reporter. Currently, creating new Reporters is undocumented,
so if you want to create your own, you'll need to make sure that you're
familiar with \href{https://adv-r.hadley.nz/R6.html}{R6} and then need read the
source code for a few.
}
\examples{
path <- testthat_example("success")

# The default reporter - doesn't display well in examples because
# it's designed to work in an interactive console.
test_file(path)

# Override the default by supplying the name of a reporter
test_file(path, reporter = "minimal")
}
\seealso{
Other reporters: 
\code{\link{CheckReporter}},
\code{\link{DebugReporter}},
\code{\link{FailReporter}},
\code{\link{ListReporter}},
\code{\link{LocationReporter}},
\code{\link{MinimalReporter}},
\code{\link{MultiReporter}},
\code{\link{ProgressReporter}},
\code{\link{RstudioReporter}},
\code{\link{SilentReporter}},
\code{\link{StopReporter}},
\code{\link{SummaryReporter}},
\code{\link{TapReporter}},
\code{\link{TeamcityReporter}}
}
\concept{reporters}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-start_reporter}{\code{Reporter$start_reporter()}}
\item \href{#method-start_context}{\code{Reporter$start_context()}}
\item \href{#method-start_test}{\code{Reporter$start_test()}}
\item \href{#method-start_file}{\code{Reporter$start_file()}}
\item \href{#method-add_result}{\code{Reporter$add_result()}}
\item \href{#method-end_test}{\code{Reporter$end_test()}}
\item \href{#method-end_context}{\code{Reporter$end_context()}}
\item \href{#method-end_reporter}{\code{Reporter$end_reporter()}}
\item \href{#method-end_file}{\code{Reporter$end_file()}}
\item \href{#method-is_full}{\code{Reporter$is_full()}}
\item \href{#method-new}{\code{Reporter$new()}}
\item \href{#method-cat_tight}{\code{Reporter$cat_tight()}}
\item \href{#method-cat_line}{\code{Reporter$cat_line()}}
\item \href{#method-rule}{\code{Reporter$rule()}}
\item \href{#method-.start_context}{\code{Reporter$.start_context()}}
\item \href{#method-.end_context}{\code{Reporter$.end_context()}}
\item \href{#method-clone}{\code{Reporter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-start_reporter"></a>}}
\subsection{Method \code{start_reporter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$start_reporter()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-start_context"></a>}}
\subsection{Method \code{start_context()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$start_context(context)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-start_test"></a>}}
\subsection{Method \code{start_test()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$start_test(context, test)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-start_file"></a>}}
\subsection{Method \code{start_file()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$start_file(filename)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_result"></a>}}
\subsection{Method \code{add_result()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$add_result(context, test, result)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-end_test"></a>}}
\subsection{Method \code{end_test()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$end_test(context, test)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-end_context"></a>}}
\subsection{Method \code{end_context()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$end_context(context)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-end_reporter"></a>}}
\subsection{Method \code{end_reporter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$end_reporter()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-end_file"></a>}}
\subsection{Method \code{end_file()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$end_file()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_full"></a>}}
\subsection{Method \code{is_full()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$is_full()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$new(file = getOption("testthat.output_file", stdout()))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cat_tight"></a>}}
\subsection{Method \code{cat_tight()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$cat_tight(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cat_line"></a>}}
\subsection{Method \code{cat_line()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$cat_line(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rule"></a>}}
\subsection{Method \code{rule()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$rule(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-.start_context"></a>}}
\subsection{Method \code{.start_context()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$.start_context(context)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-.end_context"></a>}}
\subsection{Method \code{.end_context()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$.end_context(context)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
