% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter.R
\docType{data}
\name{Reporter}
\alias{Reporter}
\title{Stub object for managing a reporter of tests.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Reporter
}
\description{
Do not clone directly from this object - children should implement all
methods.
}
\keyword{internal}
