% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-rstudio.R
\docType{data}
\name{RstudioReporter}
\alias{RstudioReporter}
\title{Test reporter: RStudio}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RstudioReporter
}
\description{
This reporter is designed for output to RStudio. It produces results in
any easily parsed form.
}
\keyword{datasets}

