% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-rstudio.R
\docType{data}
\name{RstudioReporter}
\alias{RstudioReporter}
\title{Test reporter: RStudio}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RstudioReporter
}
\description{
This reporter is designed for output to RStudio. It produces results in
any easily parsed form.
}
\seealso{
Other reporters: \code{\link{CheckReporter}},
  \code{\link{DebugReporter}}, \code{\link{FailReporter}},
  \code{\link{ListReporter}},
  \code{\link{LocationReporter}},
  \code{\link{MinimalReporter}},
  \code{\link{MultiReporter}},
  \code{\link{ProgressReporter}},
  \code{\link{SilentReporter}}, \code{\link{StopReporter}},
  \code{\link{SummaryReporter}}, \code{\link{TapReporter}},
  \code{\link{TeamcityReporter}}
}
\keyword{datasets}
