\name{setup_test_dir}
\alias{setup_test_dir}
\title{take care or finding the test files and sourcing the helpers}
\usage{
setup_test_dir(path, filter, env)
}
\arguments{
  \item{path}{path to tests}

  \item{filter}{cf \code{\link{test_dir}}}

  \item{env}{environment in which to source the helpers}
}
\value{
the test file paths
}
\description{
take care or finding the test files and sourcing the
helpers
}

