% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-silent.R
\name{SilentReporter}
\alias{SilentReporter}
\title{Test reporter: gather all errors silently.}
\description{
This reporter quietly runs all tests, simply gathering all expectations.
This is helpful for programmatically inspecting errors after a test run.
You can retrieve the results with the \code{expectations()}
method.
}
\seealso{
Other reporters: 
\code{\link{CheckReporter}},
\code{\link{DebugReporter}},
\code{\link{FailReporter}},
\code{\link{ListReporter}},
\code{\link{LocationReporter}},
\code{\link{MinimalReporter}},
\code{\link{MultiReporter}},
\code{\link{ProgressReporter}},
\code{\link{Reporter}},
\code{\link{RstudioReporter}},
\code{\link{StopReporter}},
\code{\link{SummaryReporter}},
\code{\link{TapReporter}},
\code{\link{TeamcityReporter}}
}
\concept{reporters}
\section{Super class}{
\code{\link[testthat:Reporter]{testthat::Reporter}} -> \code{SilentReporter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SilentReporter$new()}}
\item \href{#method-add_result}{\code{SilentReporter$add_result()}}
\item \href{#method-expectations}{\code{SilentReporter$expectations()}}
\item \href{#method-clone}{\code{SilentReporter$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id=".end_context">}\href{../../testthat/html/Reporter.html#method-.end_context}{\code{testthat::Reporter$.end_context()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id=".start_context">}\href{../../testthat/html/Reporter.html#method-.start_context}{\code{testthat::Reporter$.start_context()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="cat_line">}\href{../../testthat/html/Reporter.html#method-cat_line}{\code{testthat::Reporter$cat_line()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="cat_tight">}\href{../../testthat/html/Reporter.html#method-cat_tight}{\code{testthat::Reporter$cat_tight()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="end_context">}\href{../../testthat/html/Reporter.html#method-end_context}{\code{testthat::Reporter$end_context()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="end_file">}\href{../../testthat/html/Reporter.html#method-end_file}{\code{testthat::Reporter$end_file()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="end_reporter">}\href{../../testthat/html/Reporter.html#method-end_reporter}{\code{testthat::Reporter$end_reporter()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="end_test">}\href{../../testthat/html/Reporter.html#method-end_test}{\code{testthat::Reporter$end_test()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="is_full">}\href{../../testthat/html/Reporter.html#method-is_full}{\code{testthat::Reporter$is_full()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="rule">}\href{../../testthat/html/Reporter.html#method-rule}{\code{testthat::Reporter$rule()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="start_context">}\href{../../testthat/html/Reporter.html#method-start_context}{\code{testthat::Reporter$start_context()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="start_file">}\href{../../testthat/html/Reporter.html#method-start_file}{\code{testthat::Reporter$start_file()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="start_reporter">}\href{../../testthat/html/Reporter.html#method-start_reporter}{\code{testthat::Reporter$start_reporter()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="start_test">}\href{../../testthat/html/Reporter.html#method-start_test}{\code{testthat::Reporter$start_test()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SilentReporter$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_result"></a>}}
\subsection{Method \code{add_result()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SilentReporter$add_result(context, test, result)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-expectations"></a>}}
\subsection{Method \code{expectations()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SilentReporter$expectations()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SilentReporter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
