\name{SilentReporter}
\alias{SilentReporter}
\alias{SilentReporter-class}
\title{Test reporter: gather all errors silently.}
\usage{
  SilentReporter(...)
}
\arguments{
  \item{...}{Arguments used to initialise class}
}
\description{
  This reporter quietly runs all tests, simply gathering
  the results for later use. This is helpful for
  programmatically inspecting errors after a test run.
}
\keyword{debugging}

