% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-silent.R
\docType{data}
\name{SilentReporter}
\alias{SilentReporter}
\title{Test reporter: gather all errors silently.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SilentReporter
}
\description{
This reporter quietly runs all tests, simply gathering all expectations.
This is helpful for programmatically inspecting errors after a test run.
You can retrieve the results with the \code{expectations()}
method.
}
\keyword{datasets}

