% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-silent.R
\docType{data}
\name{SilentReporter}
\alias{SilentReporter}
\title{Test reporter: gather all errors silently.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SilentReporter
}
\description{
This reporter quietly runs all tests, simply gathering all expectations.
This is helpful for programmatically inspecting errors after a test run.
You can retrieve the results with the \code{expectations()}
method.
}
\seealso{
Other reporters: \code{\link{CheckReporter}},
  \code{\link{DebugReporter}}, \code{\link{FailReporter}},
  \code{\link{ListReporter}},
  \code{\link{LocationReporter}},
  \code{\link{MinimalReporter}},
  \code{\link{MultiReporter}},
  \code{\link{RstudioReporter}},
  \code{\link{StopReporter}},
  \code{\link{SummaryReporter}}, \code{\link{TapReporter}},
  \code{\link{TeamcityReporter}}
}
\keyword{datasets}
