% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skip.R
\name{skip}
\alias{skip}
\alias{skip_if_not}
\alias{skip_if_not_installed}
\alias{skip_on_appveyor}
\alias{skip_on_bioc}
\alias{skip_on_cran}
\alias{skip_on_os}
\alias{skip_on_travis}
\title{Skip a test.}
\usage{
skip(message)

skip_if_not(condition, message = deparse(substitute(condition)))

skip_if_not_installed(pkg)

skip_on_cran()

skip_on_os(os)

skip_on_travis()

skip_on_appveyor()

skip_on_bioc()
}
\arguments{
\item{message}{A message describing why the test was skipped.}

\item{condition}{Boolean condition to check. If not \code{TRUE}, will
skip the test.}

\item{pkg}{Name of package to check for}

\item{os}{Character vector of system names. Supported values are
\code{"windows"}, \code{"mac"}, \code{"linux"} and \code{"solaris"}.}
}
\description{
This function allows you to skip a test if it's not currently available.
This will produce an informative message, but will not cause the test
suite to fail.
}
\section{Helpers}{

\code{skip_if_not()} works like \code{\link{stopifnot}}, generating
a message automatically based on the first argument.

\code{skip_on_cran()} skips tests on CRAN, using the \code{NOT_CRAN}
environment variable set by devtools.

\code{skip_on_travis()} skips tests on travis by inspecting the
\code{TRAVIS} environment variable.

\code{skip_on_appveyor()} skips tests on appveyor by inspecting the
\code{APPVEYOR} environment variable.

\code{skip_on_bioc()} skips tests on Bioconductor by inspecting the
\code{BBS_HOME} environment variable.

\code{skip_if_not_installed()} skips a tests if a package is not installed
(useful for suggested packages).
}
\examples{
if (FALSE) skip("No internet connection")
}

