% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot-manage.R
\name{snapshot_accept}
\alias{snapshot_accept}
\alias{snapshot_review}
\title{Snapshot management}
\usage{
snapshot_accept(files = NULL, path = "tests/testthat")

snapshot_review(files = NULL, path = "tests/testthat")
}
\arguments{
\item{files}{Optionally, filter affects to snapshots from specified test
files. Can be full path to test (\code{tests/testthat/test-foo.R}), file name
(\code{test-foo.R}), or test name (\code{foo}).}

\item{path}{Path to tests.}
}
\description{
\itemize{
\item \code{snapshot_accept()} accepts all modified snapshots.
\item \code{snapshot_review()} opens a Shiny app that shows a visual diff of each
modified snapshot. This is particularly useful for whole file snapshots
created by \code{expect_snapshot_file()}.
}
}
