% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot-github.R
\name{snapshot_download_gh}
\alias{snapshot_download_gh}
\title{Download snapshots from GitHub}
\usage{
snapshot_download_gh(repository, run_id, dest_dir = ".")
}
\arguments{
\item{repository}{Repository name, e.g. \code{"r-lib/testthat"}.}

\item{run_id}{Run ID, e.g. \code{"47905180716"}. You can find this in the action url.}

\item{dest_dir}{Directory to download to. Defaults to the current directory.}
}
\description{
If your snapshots fail on GitHub, it can be a pain to figure out exactly
why, or to incporate them into your local package. This function makes it
easy.

Note that you should not generally need to fill out this function yourself;
instead copy and paste from the hint emitted on GitHub.
}
