% Generated by roxygen2 (4.0.2): do not edit by hand
\name{source_dir}
\alias{source_dir}
\title{Load all source files in a directory.}
\usage{
source_dir(path, pattern = "\\\\.[rR]$", env = test_env(), chdir = TRUE)
}
\arguments{
\item{path}{path to tests}

\item{pattern}{regular expression used to filter files}

\item{env}{environment in which to store results}

\item{chdir}{change working directory to path?}
}
\description{
The expectation is that the files can be sourced in alphabetical order.
}
\keyword{internal}

