% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-stop.R
\name{StopReporter}
\alias{StopReporter}
\title{Error if any test fails}
\description{
The default reporter used when \code{\link[=expect_that]{expect_that()}} is run interactively.
It responds by displaying a summary of the number of successes and failures
and \code{\link[=stop]{stop()}}ping on if there are any failures.
}
\seealso{
Other reporters:
\code{\link{CheckReporter}},
\code{\link{DebugReporter}},
\code{\link{FailReporter}},
\code{\link{JunitReporter}},
\code{\link{ListReporter}},
\code{\link{LlmReporter}},
\code{\link{LocationReporter}},
\code{\link{MinimalReporter}},
\code{\link{MultiReporter}},
\code{\link{ProgressReporter}},
\code{\link{RStudioReporter}},
\code{\link{Reporter}},
\code{\link{SilentReporter}},
\code{\link{SlowReporter}},
\code{\link{SummaryReporter}},
\code{\link{TapReporter}},
\code{\link{TeamcityReporter}}
}
\concept{reporters}
