% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-stop.R
\docType{data}
\name{StopReporter}
\alias{StopReporter}
\title{Test reporter: stop on error.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
StopReporter
}
\description{
The default reporter, executed when \code{expect_that} is run
interactively. It
responds by \link{stop}()ing on failures and doing nothing otherwise. This
will ensure that a failing test will raise an error.
}
\details{
This should be used when doing a quick and dirty test, or during the final
automated testing of R CMD check.  Otherwise, use a reporter that runs all
tests and gives you more context about the problem.
}
\keyword{datasets}

