% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect-that.R
\name{succeed}
\alias{succeed}
\title{Mark a test as successful}
\usage{
succeed(message = "Success has been forced", info = NULL)
}
\arguments{
\item{message}{a string to display.}

\item{info}{Character vector continuing additional information. Included
for backward compatibility only and new expectations should not use it.}
}
\description{
This is an older version of \code{\link[=pass]{pass()}} that exists for backwards compatibility.
You should now use \code{pass()} instead`
}
\keyword{internal}
