% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expect-that.r
\name{succeed}
\alias{succeed}
\title{A default expectation that always succeeds.}
\usage{
succeed(message = "Success has been forced")
}
\arguments{
\item{message}{a string to display.}
}
\description{
A default expectation that always succeeds.
}
\examples{
\dontrun{
test_that("this test fails", fail())
}
}

