% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-summary.R
\name{SummaryReporter}
\alias{SummaryReporter}
\title{Report a summary of failures}
\description{
This is designed for interactive usage: it lets you know which tests have
run successfully and as well as fully reporting information about
failures and errors.

You can use the \code{max_reports} field to control the maximum number
of detailed reports produced by this reporter.

As an additional benefit, this reporter will praise you from time-to-time
if all your tests pass.
}
\seealso{
Other reporters: 
\code{\link{CheckReporter}},
\code{\link{DebugReporter}},
\code{\link{FailReporter}},
\code{\link{JunitReporter}},
\code{\link{ListReporter}},
\code{\link{LlmReporter}},
\code{\link{LocationReporter}},
\code{\link{MinimalReporter}},
\code{\link{MultiReporter}},
\code{\link{ProgressReporter}},
\code{\link{RStudioReporter}},
\code{\link{Reporter}},
\code{\link{SilentReporter}},
\code{\link{SlowReporter}},
\code{\link{SonarqubeReporter}},
\code{\link{StopReporter}},
\code{\link{TapReporter}},
\code{\link{TeamcityReporter}}
}
\concept{reporters}
