\name{SummaryReporter}
\alias{SummaryReporter}
\title{Test reporter: summary of errors.}

\description{
  Test reporter: summary of errors.
}

\details{
  This is the most useful reporting reporter as it lets you
  know both which tests have run successfully, as well as
  fully reporting information about failures and errors.
  It is the default reporting reporter used by
  \code{\link{test_dir}} and \code{\link{test_file}}.

  As an additional benefit, this reporter will praise you
  from time-to-time if all your tests pass.
}
\keyword{debugging}
