% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectations.r
\name{takes_less_than}
\alias{takes_less_than}
\title{Expectation: does expression take less than a fixed amount of time to run?}
\usage{
takes_less_than(amount)
}
\arguments{
\item{amount}{maximum duration in seconds}
}
\description{
This is useful for performance regression testing.
}
\seealso{
Other expectations: \code{\link{equivalence}},
  \code{\link{expect-compare}},
  \code{\link{expect_equal_to_reference}},
  \code{\link{expect_is}}, \code{\link{expect_named}},
  \code{\link{expect_silent}}, \code{\link{expect_true}},
  \code{\link{matching-expectations}}
}
\keyword{internal}

