\name{takes_less_than}
\alias{takes_less_than}
\title{Expectation: does expression take less than a fixed amount of time to run?}
\usage{
  takes_less_than(amount)
}
\arguments{
  \item{amount}{maximum duration in seconds}
}
\description{
  This is useful for performance regression testing.
}
\seealso{
  Other expectations: \code{\link{equals}},
  \code{\link{expect_equal}},
  \code{\link{expect_equivalent}},
  \code{\link{expect_error}}, \code{\link{expect_false}},
  \code{\link{expect_identical}}, \code{\link{expect_is}},
  \code{\link{expect_match}}, \code{\link{expect_message}},
  \code{\link{expect_output}}, \code{\link{expect_true}},
  \code{\link{expect_warning}},
  \code{\link{gives_warning}}, \code{\link{is_a}},
  \code{\link{is_equivalent_to}}, \code{\link{is_false}},
  \code{\link{is_identical_to}}, \code{\link{is_true}},
  \code{\link{matches}}, \code{\link{prints_text}},
  \code{\link{shows_message}}, \code{\link{throws_error}}
}

