\name{takes_less_than}
\alias{takes_less_than}
\title{Expectation: does expression take less than a fixed amount of time to run?}
\usage{takes_less_than(amount)}

\description{
  Expectation: does expression take less than a fixed
  amount of time to run?
}

\details{
  This is useful for performance regression testing.
}
\arguments{
  \item{amount}{maximum duration in seconds}
}
