% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old-school.R
\name{takes_less_than}
\alias{takes_less_than}
\title{Does code take less than the expected amount of time to run?}
\usage{
takes_less_than(amount)
}
\arguments{
\item{amount}{maximum duration in seconds}
}
\description{
This is useful for performance regression testing.
}
\keyword{internal}
