\name{TapReporter}
\alias{TapReporter}
\alias{TapReporter-class}
\title{Test reporter: TAP format.}
\description{
  This reporter will output results in the Test Anything
  Protocol (TAP), a simple text-based interface between
  testing modules in a test harness. For more information
  about TAP, see http://testanything.org
}
\keyword{debugging}

