% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-tap.R
\docType{data}
\name{TapReporter}
\alias{TapReporter}
\title{Test reporter: TAP format.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TapReporter
}
\description{
This reporter will output results in the Test Anything Protocol (TAP),
a simple text-based interface between testing modules in a test harness.
For more information about TAP, see http://testanything.org
}
\keyword{datasets}

