% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-tap.R
\name{TapReporter}
\alias{TapReporter}
\title{Test reporter: TAP format.}
\description{
This reporter will output results in the Test Anything Protocol (TAP),
a simple text-based interface between testing modules in a test harness.
For more information about TAP, see http://testanything.org
}
\seealso{
Other reporters: 
\code{\link{CheckReporter}},
\code{\link{DebugReporter}},
\code{\link{FailReporter}},
\code{\link{JunitReporter}},
\code{\link{ListReporter}},
\code{\link{LocationReporter}},
\code{\link{MinimalReporter}},
\code{\link{MultiReporter}},
\code{\link{ProgressReporter}},
\code{\link{RStudioReporter}},
\code{\link{Reporter}},
\code{\link{SilentReporter}},
\code{\link{StopReporter}},
\code{\link{SummaryReporter}},
\code{\link{TeamcityReporter}}
}
\concept{reporters}
