% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{TeamcityReporter-class}
\alias{TeamcityReporter}
\alias{TeamcityReporter-class}
\title{Test reporter: Teamcity format.}
\arguments{
\item{...}{Arguments used to initialise class}
}
\description{
This reporter will output results in the Teamcity message format.
For more information about Teamcity messages, see
http://confluence.jetbrains.com/display/TCD7/Build+Script+Interaction+with+TeamCity
}
\keyword{debugging}

