% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-teamcity.R
\docType{data}
\name{TeamcityReporter}
\alias{TeamcityReporter}
\title{Test reporter: Teamcity format.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TeamcityReporter
}
\description{
This reporter will output results in the Teamcity message format.
For more information about Teamcity messages, see
http://confluence.jetbrains.com/display/TCD7/Build+Script+Interaction+with+TeamCity
}
\keyword{datasets}

