% Generated by roxygen2 (4.0.2): do not edit by hand
\name{test_dir}
\alias{test_dir}
\title{Run all of the tests in a directory.}
\usage{
test_dir(path, filter = NULL, reporter = "summary", env = test_env())
}
\arguments{
\item{path}{path to tests}

\item{filter}{If not \code{NULL}, only tests with file names matching this
regular expression will be executed.  Matching will take on the file
name after it has been stripped of \code{"test-"} and \code{".r"}.}

\item{reporter}{reporter to use}

\item{env}{environment in which to execute test suite.}
}
\value{
a data frame of the summary of test results
}
\description{
Test files start with \code{test} and are executed in alphabetical order
(but they shouldn't have dependencies). Helper files start with
\code{helper} and loaded before any tests are run.
}

