% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-files.R
\name{test_dir}
\alias{test_dir}
\title{Run all of the tests in a directory.}
\usage{
test_dir(path, filter = NULL, reporter = "summary", env = test_env(), ...,
  load_helpers = TRUE)
}
\arguments{
\item{path}{path to tests}

\item{filter}{If not \code{NULL}, only tests with file names matching this
regular expression will be executed.  Matching will take on the file
name after it has been stripped of \code{"test-"} and \code{".R"}.}

\item{reporter}{reporter to use}

\item{env}{environment in which to execute the tests}

\item{...}{Additional arguments passed to \code{grepl} to control filtering.}

\item{load_helpers}{Source helper files before running the tests?}
}
\value{
the results as a "testthat_results" (list)
}
\description{
Test files start with \code{test} and are executed in alphabetical order
(but they shouldn't have dependencies). Helper files start with
\code{helper} and loaded before any tests are run.
}

