% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-files.R
\name{test_dir}
\alias{test_dir}
\title{Run all of the tests in a directory.}
\usage{
test_dir(path, filter = NULL, reporter = default_reporter(),
  env = test_env(), ..., encoding = "unknown", load_helpers = TRUE)
}
\arguments{
\item{path}{path to tests}

\item{filter}{If not \code{NULL}, only tests with file names matching this
regular expression will be executed.  Matching will take on the file
name after it has been stripped of \code{"test-"} and \code{".R"}.}

\item{reporter}{reporter to use}

\item{env}{environment in which to execute the tests}

\item{...}{Additional arguments passed to \code{\link[=grepl]{grepl()}} to control filtering.}

\item{encoding}{File encoding, default is "unknown"
\code{unknown}.}

\item{load_helpers}{Source helper files before running the tests?}
}
\value{
The results of the reporter function on all test results.
}
\description{
Test files start with \code{test} and are executed in alphabetical order
(but they shouldn't have dependencies). Helper files start with
\code{helper} and loaded before any tests are run.
}
