% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-env.R
\name{test_env}
\alias{test_env}
\title{Generate default testing environment.}
\usage{
test_env(package = NULL)
}
\description{
We use a new environment which inherits from \code{\link[=globalenv]{globalenv()}} or a package
namespace. In an ideal world, we'd avoid putting the global environment on
the search path for tests, but it's not currently possible without losing
the ability to load packages in tests.
}
\keyword{internal}
