% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-files.R
\name{test_file}
\alias{test_file}
\title{Run all tests in a single file}
\usage{
test_file(path, reporter = default_compact_reporter(), ...)
}
\arguments{
\item{path}{Path to file.}

\item{reporter}{Reporter to use to summarise output. Can be supplied
as a string (e.g. "summary") or as an R6 object
(e.g. \code{SummaryReporter$new()}).

See \link{Reporter} for more details and a list of built-in reporters.}

\item{...}{Additional parameters passed on to \code{test_dir()}}
}
\value{
Invisibly, a list with one element for each test.
}
\description{
A variant of \code{\link[=test_dir]{test_dir()}} that only tests a single file within a directory.
The directory is still used to locate helpers, setup and teardown files.
}
\examples{
path <- testthat_example("success")
test_file(path)
test_file(path, reporter = "minimal")
}
