% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-files.R
\name{test_file}
\alias{test_file}
\title{Run all tests in specified file.}
\usage{
test_file(path, reporter = default_reporter(), env = test_env(),
  start_end_reporter = TRUE, load_helpers = TRUE, encoding = "unknown",
  wrap = TRUE)
}
\arguments{
\item{path}{path to file}

\item{reporter}{reporter to use}

\item{env}{environment in which to execute the tests}

\item{start_end_reporter}{whether to start and end the reporter}

\item{load_helpers}{Source helper files before running the tests?}

\item{encoding}{File encoding, default is "unknown"
\code{unknown}.}

\item{wrap}{Automatically wrap all code within \code{\link[=test_that]{test_that()}}? This ensures
that all expectations are reported, even if outside a test block.}
}
\value{
the results as a "testthat_results" (list)
}
\description{
Run all tests in specified file.
}
