% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test-files.r
\name{test_file}
\alias{test_file}
\title{Run all tests in specified file.}
\usage{
test_file(path, reporter = "summary", env = test_env(),
  start_end_reporter = TRUE)
}
\arguments{
\item{path}{path to file}

\item{reporter}{reporter to use}

\item{env}{environment in which to execute the tests}

\item{start_end_reporter}{whether to start and end the reporter}
}
\value{
the results as a "testthat_results" (list)
}
\description{
Run all tests in specified file.
}

