% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-package.R
\name{test_package}
\alias{test_package}
\alias{test_check}
\title{Run all tests in an installed package.}
\usage{
test_package(package, filter = NULL, reporter = default_reporter(), ...,
  stop_on_failure = TRUE, stop_on_warning = FALSE)

test_check(package, filter = NULL,
  reporter = getOption("testthat.default_check_reporter", "check"), ...,
  stop_on_failure = TRUE, stop_on_warning = FALSE)
}
\arguments{
\item{package}{package name}

\item{filter}{If not \code{NULL}, only tests with file names matching this
regular expression will be executed.  Matching will take on the file
name after it has been stripped of \code{"test-"} and \code{".R"}.}

\item{reporter}{reporter to use}

\item{...}{Additional arguments passed to \code{\link[=grepl]{grepl()}} to control filtering.}

\item{stop_on_failure}{If \code{TRUE}, throw an error if any tests fail.}

\item{stop_on_warning}{If \code{TRUE}, throw an error if any tests generate
warnings.}
}
\value{
the results as a "testthat_results" (list)
}
\description{
Test are run in an environment that inherits from the package's namespace
environment, so that tests can access non-exported functions and variables.
Tests should be placed in \code{tests/testthat}. Use \code{test_check()} with
\code{R CMD check} and \code{test_package()} interactively at the console.
}
\section{R CMD check}{

Create \code{tests/testthat.R} that contains:

\preformatted{
library(testthat)
library(yourpackage)

test_check("yourpackage")
}
}

\examples{
\dontrun{test_package("testthat")}
}
