% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-path.R
\name{test_path}
\alias{test_path}
\title{Locate file in testing directory.}
\usage{
test_path(...)
}
\arguments{
\item{...}{Character vectors giving path component.}
}
\value{
A character vector giving the path.
}
\description{
This function is designed to work both interactively and during tests,
locating files in the \code{tests/testthat} directory
}
