% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test-that.r
\name{test_that}
\alias{test_that}
\title{Create a test.}
\usage{
test_that(desc, code, suppress = getOption("testthat.suppress", c("messages",
  "warnings")))
}
\arguments{
\item{desc}{test name.  Names should be kept as brief as possible, as they
are often used as line prefixes.}

\item{code}{test code containing expectations}

\item{suppress}{which conditions to suppress}
}
\description{
A test encapsulates a series of expectations about small, self-contained
set of functionality.  Each test is contained in a \link{context} and
contains multiple expectation generated by \code{\link{equivalence}}.
}
\details{
Tests are evaluated in their own environments, and should not affect
global state.

When run from the command line, tests return \code{NULL} if all
expectations are met, otherwise it raises an error.
}
\examples{
test_that("trigonometric functions match identities", {
  expect_equal(sin(pi / 4), 1 / sqrt(2))
  expect_equal(cos(pi / 4), 1 / sqrt(2))
  expect_equal(tan(pi / 4), 1)
})
# Failing test:
\dontrun{
test_that("trigonometric functions match identities", {
  expect_equal(sin(pi / 4), 1)
})
}
# Turning off suppression of warnings
test_that("This warning is not suppressed", {
  warning("test")
}, suppress = "messages")
# Turn off all suppression
test_that("This warning and message are not suppressed", {
  warning("a warning!")
  message("a message!")
}, suppress = NULL)
}

