% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-that.R
\docType{package}
\name{testthat-package}
\alias{testthat}
\alias{testthat-package}
\title{R package to make testing fun!}
\description{
Try the example below. Have a look at the references and learn more
from function documentation such as \code{\link[=expect_that]{expect_that()}}.
}
\section{Options}{

\itemize{
\item \code{testthat.use_colours}: Should the output be coloured? (Default: \code{TRUE}).
\item \code{testthat.summary.max_reports}: The maximum number of detailed test
reports printed for the summary reporter (default: 10).
\item \code{testthat.summary.omit_dots}: Omit progress dots in the summary reporter
(default: \code{FALSE}).
}
}

\examples{
library(testthat)
a <- 9
expect_that(a, is_less_than(10))
expect_lt(a, 10)
}
\references{
Wickham, H (2011). testthat: Get Started with Testing.
\strong{The R Journal} \emph{3/1} 5-10.
\url{http://journal.r-project.org/archive/2011-1/RJournal_2011-1_Wickham.pdf}

\url{http://adv-r.had.co.nz/Testing.html}
}
\seealso{
Useful links:
\itemize{
  \item \url{http://testthat.r-lib.org}
  \item \url{https://github.com/r-lib/testthat}
  \item Report bugs at \url{https://github.com/r-lib/testthat/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
  \item R Core team (Implementation of utils::recover()) [contributor]
}

}
\keyword{internal}
