% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test-results.r
\name{testthat_results}
\alias{testthat_results}
\title{Create a `testthat_results` object from the test results
as stored in the ListReporter results field.}
\usage{
testthat_results(results)
}
\arguments{
\item{results}{a list as stored in ListReporter}
}
\value{
its list argument as a `testthat_results` object
}
\description{
Create a `testthat_results` object from the test results
as stored in the ListReporter results field.
}
\seealso{
ListReporter
}

