% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test-compiled-code.r
\name{use_catch}
\alias{use_catch}
\title{Use Catch for C++ Unit Testing}
\usage{
use_catch(dir = getwd())
}
\arguments{
\item{dir}{The directory containing an \R package.}
}
\description{
Add the necessary infrastructure to enable C++ unit testing
in \R packages with
\href{https://github.com/philsquared/Catch}{Catch} and \code{testthat}.
}
\details{
This function will:

\enumerate{
  \item Create a file \code{src/test-runner.cpp},
  \item Ensure that your package links to \code{testthat}.
}

C++ unit tests can be added to C++ source files within the
\code{src/} directory of your package, with a format similar
to \R code tested with \code{testthat} -- for example,

\preformatted{
context("C++ Unit Test") {
  test_that("two plus two is four") {
    int result = 2 + 2;
    expect_true(result == 4);
  }
}
}

When your package is compiled, unit tests alongside a harness
for running these tests will be compiled into the \R package,
with the entry point \code{run_testthat_tests()}. \code{testthat}
will use that entry point to run your unit tests when detected.
}
\seealso{
\href{https://github.com/philsquared/Catch}{Catch}, the
  library used to enable C++ unit testing.
}

