% Generated by roxygen2 (4.0.2): do not edit by hand
\name{with_mock}
\alias{with_mock}
\title{Mock functions in a package.}
\usage{
with_mock(..., .env = topenv())
}
\arguments{
\item{...}{named parameters redefine mocked functions, unnamed parameters
will be evaluated after mocking the functions}

\item{.env}{the environment in which to patch the functions,
defaults to the top-level environment.  A character is interpreted as
package name.}
}
\value{
The result of the last unnamed parameter
}
\description{
Executes code after temporarily substituting implementations of package
functions.  This is useful for testing code that relies on functions that are
slow, have unintended side effects or access resources that may not be
available when testing.
}
\examples{
with_mock(
  `base::all.equal` = function(x, y, ...) TRUE,
  {
    expect_equal(3, 5)
  }
)
throws_error()(expect_equal(3, 5))
}

